Refactor database models
models.py was too big for Claude, so I refactored
it into a models/ module.
