--- app/backend/src/couchers/models.py
@@ -1,2987 +1,0 @@
-import enum
-
-from geoalchemy2.types import Geometry
-from google.protobuf import empty_pb2
-from sqlalchemy import (
-    ARRAY,
-    JSON,
-    BigInteger,
-    Boolean,
-    CheckConstraint,
-    Column,
-    Date,
-    DateTime,
-    Enum,
-    Float,
-    ForeignKey,
-    Index,
-    Integer,
-    Interval,
-    MetaData,
-    Sequence,
-    String,
-    UniqueConstraint,
-)
-from sqlalchemy import LargeBinary as Binary
-from sqlalchemy.dialects.postgresql import INET, TSTZRANGE, ExcludeConstraint
-from sqlalchemy.ext.associationproxy import association_proxy
-from sqlalchemy.ext.hybrid import hybrid_method, hybrid_property
-from sqlalchemy.orm import backref, column_property, declarative_base, deferred, relationship
-from sqlalchemy.sql import and_, expression, func, not_, or_, text
-from sqlalchemy.sql import select as sa_select
-
-from couchers import urls
-from couchers.config import config
-from couchers.constants import (
-    DATETIME_INFINITY,
-    DATETIME_MINUS_INFINITY,
-    EMAIL_REGEX,
-    GUIDELINES_VERSION,
-    PHONE_VERIFICATION_LIFETIME,
-    SMS_CODE_LIFETIME,
-    TOS_VERSION,
-)
-from couchers.utils import (
-    date_in_timezone,
-    get_coordinates,
-    last_active_coarsen,
-    now,
-)
-from proto import notification_data_pb2
-
-meta = MetaData(
-    naming_convention={
-        "ix": "ix_%(column_0_label)s",
-        "uq": "uq_%(table_name)s_%(column_0_name)s",
-        "ck": "ck_%(table_name)s_%(constraint_name)s",
-        "fk": "fk_%(table_name)s_%(column_0_name)s_%(referred_table_name)s",
-        "pk": "pk_%(table_name)s",
-    }
-)
-
-Base = declarative_base(metadata=meta)
-
-
-class HostingStatus(enum.Enum):
-    can_host = enum.auto()
-    maybe = enum.auto()
-    cant_host = enum.auto()
-
-
-class MeetupStatus(enum.Enum):
-    wants_to_meetup = enum.auto()
-    open_to_meetup = enum.auto()
-    does_not_want_to_meetup = enum.auto()
-
-
-class SmokingLocation(enum.Enum):
-    yes = enum.auto()
-    window = enum.auto()
-    outside = enum.auto()
-    no = enum.auto()
-
-
-class SleepingArrangement(enum.Enum):
-    private = enum.auto()
-    common = enum.auto()
-    shared_room = enum.auto()
-
-
-class ParkingDetails(enum.Enum):
-    free_onsite = enum.auto()
-    free_offsite = enum.auto()
-    paid_onsite = enum.auto()
-    paid_offsite = enum.auto()
-
-
-class ProfilePublicVisibility(enum.Enum):
-    # no public info
-    nothing = enum.auto()
-    # only show on map, randomized, unclickable
-    map_only = enum.auto()
-    # name, gender, location, hosting/meetup status, badges, number of references, and signup time
-    limited = enum.auto()
-    # full about me except additional info (hide my home)
-    most = enum.auto()
-    # all but references
-    full = enum.auto()
-
-
-class TimezoneArea(Base):
-    __tablename__ = "timezone_areas"
-    id = Column(BigInteger, primary_key=True)
-
-    tzid = Column(String)
-    geom = Column(Geometry(geometry_type="MULTIPOLYGON", srid=4326), nullable=False)
-
-    __table_args__ = (
-        Index(
-            "ix_timezone_areas_geom_tzid",
-            geom,
-            tzid,
-            postgresql_using="gist",
-        ),
-    )
-
-
-class User(Base):
-    """
-    Basic user and profile details
-    """
-
-    __tablename__ = "users"
-
-    id = Column(BigInteger, primary_key=True)
-
-    username = Column(String, nullable=False, unique=True)
-    email = Column(String, nullable=False, unique=True)
-    # stored in libsodium hash format, can be null for email login
-    hashed_password = Column(Binary, nullable=False)
-    # phone number in E.164 format with leading +, for example "+46701740605"
-    phone = Column(String, nullable=True, server_default=expression.null())
-    # language preference -- defaults to empty string
-    ui_language_preference = Column(String, nullable=True, server_default="")
-
-    # timezones should always be UTC
-    ## location
-    # point describing their location. EPSG4326 is the SRS (spatial ref system, = way to describe a point on earth) used
-    # by GPS, it has the WGS84 geoid with lat/lon
-    geom = Column(Geometry(geometry_type="POINT", srid=4326), nullable=False)
-    # randomized coordinates within a radius of 0.02-0.1 degrees, equates to about 2-10 km
-    randomized_geom = Column(Geometry(geometry_type="POINT", srid=4326), nullable=True)
-    # their display location (displayed to other users), in meters
-    geom_radius = Column(Float, nullable=False)
-    # the display address (text) shown on their profile
-    city = Column(String, nullable=False)
-    # "Grew up in" on profile
-    hometown = Column(String, nullable=True)
-
-    regions_visited = relationship("Region", secondary="regions_visited", order_by="Region.name")
-    regions_lived = relationship("Region", secondary="regions_lived", order_by="Region.name")
-
-    timezone = column_property(
-        sa_select(TimezoneArea.tzid).where(func.ST_Contains(TimezoneArea.geom, geom)).limit(1).scalar_subquery(),
-        deferred=True,
-    )
-
-    joined = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    last_active = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    public_visibility = Column(Enum(ProfilePublicVisibility), nullable=False, server_default="map_only")
-    has_modified_public_visibility = Column(Boolean, nullable=False, server_default=expression.false())
-
-    # id of the last message that they received a notification about
-    last_notified_message_id = Column(BigInteger, nullable=False, default=0)
-    # same as above for host requests
-    last_notified_request_message_id = Column(BigInteger, nullable=False, server_default=text("0"))
-
-    # display name
-    name = Column(String, nullable=False)
-    gender = Column(String, nullable=False)
-    pronouns = Column(String, nullable=True)
-    birthdate = Column(Date, nullable=False)  # in the timezone of birthplace
-
-    avatar_key = Column(ForeignKey("uploads.key"), nullable=True)
-
-    hosting_status = Column(Enum(HostingStatus), nullable=False)
-    meetup_status = Column(Enum(MeetupStatus), nullable=False, server_default="open_to_meetup")
-
-    # community standing score
-    community_standing = Column(Float, nullable=True)
-
-    occupation = Column(String, nullable=True)  # CommonMark without images
-    education = Column(String, nullable=True)  # CommonMark without images
-
-    # "Who I am" under "About Me" tab
-    about_me = Column(String, nullable=True)  # CommonMark without images
-    # "What I do in my free time" under "About Me" tab
-    things_i_like = Column(String, nullable=True)  # CommonMark without images
-    # "About my home" under "My Home" tab
-    about_place = Column(String, nullable=True)  # CommonMark without images
-    # "Additional information" under "About Me" tab
-    additional_information = Column(String, nullable=True)  # CommonMark without images
-
-    is_banned = Column(Boolean, nullable=False, server_default=expression.false())
-    is_deleted = Column(Boolean, nullable=False, server_default=expression.false())
-    is_superuser = Column(Boolean, nullable=False, server_default=expression.false())
-
-    # the undelete token allows a user to recover their account for a couple of days after deletion in case it was
-    # accidental or they changed their mind
-    # constraints make sure these are non-null only if is_deleted and that these are null in unison
-    undelete_token = Column(String, nullable=True)
-    # validity of the undelete token
-    undelete_until = Column(DateTime(timezone=True), nullable=True)
-
-    # hosting preferences
-    max_guests = Column(Integer, nullable=True)
-    last_minute = Column(Boolean, nullable=True)
-    has_pets = Column(Boolean, nullable=True)
-    accepts_pets = Column(Boolean, nullable=True)
-    pet_details = Column(String, nullable=True)  # CommonMark without images
-    has_kids = Column(Boolean, nullable=True)
-    accepts_kids = Column(Boolean, nullable=True)
-    kid_details = Column(String, nullable=True)  # CommonMark without images
-    has_housemates = Column(Boolean, nullable=True)
-    housemate_details = Column(String, nullable=True)  # CommonMark without images
-    wheelchair_accessible = Column(Boolean, nullable=True)
-    smoking_allowed = Column(Enum(SmokingLocation), nullable=True)
-    smokes_at_home = Column(Boolean, nullable=True)
-    drinking_allowed = Column(Boolean, nullable=True)
-    drinks_at_home = Column(Boolean, nullable=True)
-    # "Additional information" under "My Home" tab
-    other_host_info = Column(String, nullable=True)  # CommonMark without images
-
-    # "Sleeping privacy" (not long-form text)
-    sleeping_arrangement = Column(Enum(SleepingArrangement), nullable=True)
-    # "Sleeping arrangement" under "My Home" tab
-    sleeping_details = Column(String, nullable=True)  # CommonMark without images
-    # "Local area information" under "My Home" tab
-    area = Column(String, nullable=True)  # CommonMark without images
-    # "House rules" under "My Home" tab
-    house_rules = Column(String, nullable=True)  # CommonMark without images
-    parking = Column(Boolean, nullable=True)
-    parking_details = Column(Enum(ParkingDetails), nullable=True)  # CommonMark without images
-    camping_ok = Column(Boolean, nullable=True)
-
-    accepted_tos = Column(Integer, nullable=False, default=0)
-    accepted_community_guidelines = Column(Integer, nullable=False, server_default="0")
-    # whether the user has yet filled in the contributor form
-    filled_contributor_form = Column(Boolean, nullable=False, server_default=expression.false())
-
-    # number of onboarding emails sent
-    onboarding_emails_sent = Column(Integer, nullable=False, server_default="0")
-    last_onboarding_email_sent = Column(DateTime(timezone=True), nullable=True)
-
-    # whether we need to sync the user's newsletter preferences with the newsletter server
-    in_sync_with_newsletter = Column(Boolean, nullable=False, server_default=expression.false())
-    # opted out of the newsletter
-    opt_out_of_newsletter = Column(Boolean, nullable=False, server_default=expression.false())
-
-    # set to null to receive no digests
-    digest_frequency = Column(Interval, nullable=True)
-    last_digest_sent = Column(DateTime(timezone=True), nullable=False, server_default=text("to_timestamp(0)"))
-
-    # for changing their email
-    new_email = Column(String, nullable=True)
-
-    new_email_token = Column(String, nullable=True)
-    new_email_token_created = Column(DateTime(timezone=True), nullable=True)
-    new_email_token_expiry = Column(DateTime(timezone=True), nullable=True)
-
-    recommendation_score = Column(Float, nullable=False, server_default="0")
-
-    # Columns for verifying their phone number. State chart:
-    #                                       ,-------------------,
-    #                                       |    Start          |
-    #                                       | phone = None      |  someone else
-    # ,-----------------,                   | token = None      |  verifies            ,-----------------------,
-    # |  Code Expired   |                   | sent = 1970 or zz |  phone xx            |  Verification Expired |
-    # | phone = xx      |  time passes      | verified = None   | <------,             | phone = xx            |
-    # | token = yy      | <------------,    | attempts = 0      |        |             | token = None          |
-    # | sent = zz (exp.)|              |    '-------------------'        |             | sent = zz             |
-    # | verified = None |              |       V    ^                    +-----------< | verified = ww (exp.)  |
-    # | attempts = 0..2 | >--,         |       |    | ChangePhone("")    |             | attempts = 0          |
-    # '-----------------'    +-------- | ------+----+--------------------+             '-----------------------'
-    #                        |         |       |    | ChangePhone(xx)    |                       ^ time passes
-    #                        |         |       ^    V                    |                       |
-    # ,-----------------,    |         |    ,-------------------,        |             ,-----------------------,
-    # |    Too Many     | >--'         '--< |    Code sent      | >------+             |         Verified      |
-    # | phone = xx      |                   | phone = xx        |        |             | phone = xx            |
-    # | token = yy      | VerifyPhone(wrong)| token = yy        |        '-----------< | token = None          |
-    # | sent = zz       | <------+--------< | sent = zz         |                      | sent = zz             |
-    # | verified = None |        |          | verified = None   | VerifyPhone(correct) | verified = ww         |
-    # | attempts = 3    |        '--------> | attempts = 0..2   | >------------------> | attempts = 0          |
-    # '-----------------'                   '-------------------'                      '-----------------------'
-
-    # randomly generated Luhn 6-digit string
-    phone_verification_token = Column(String(6), nullable=True, server_default=expression.null())
-
-    phone_verification_sent = Column(DateTime(timezone=True), nullable=False, server_default=text("to_timestamp(0)"))
-    phone_verification_verified = Column(DateTime(timezone=True), nullable=True, server_default=expression.null())
-    phone_verification_attempts = Column(Integer, nullable=False, server_default=text("0"))
-
-    # the stripe customer identifier if the user has donated to Couchers
-    # e.g. cus_JjoXHttuZopv0t
-    # for new US entity
-    stripe_customer_id = Column(String, nullable=True)
-    # for old AU entity
-    stripe_customer_id_old = Column(String, nullable=True)
-
-    has_passport_sex_gender_exception = Column(Boolean, nullable=False, server_default=expression.false())
-
-    #  checking for phone verification
-    has_donated = Column(Boolean, nullable=False, server_default=expression.false())
-
-    # whether this user has all emails turned off
-    do_not_email = Column(Boolean, nullable=False, server_default=expression.false())
-
-    avatar = relationship("Upload", foreign_keys="User.avatar_key")
-
-    admin_note = Column(String, nullable=False, server_default=text("''"))
-
-    # whether mods have marked this user has having to update their location
-    needs_to_update_location = Column(Boolean, nullable=False, server_default=expression.false())
-
-    last_antibot = Column(DateTime(timezone=True), nullable=False, server_default=text("to_timestamp(0)"))
-
-    age = column_property(func.date_part("year", func.age(birthdate)))
-
-    # ID of the invite code used to sign up (if any)
-    invite_code_id = Column(ForeignKey("invite_codes.id"), nullable=True)
-    invite_code = relationship("InviteCode", foreign_keys=[invite_code_id])
-
-    moderation_user_lists = relationship(
-        "ModerationUserList", secondary="moderation_user_list_members", back_populates="users"
-    )
-
-    __table_args__ = (
-        # Verified phone numbers should be unique
-        Index(
-            "ix_users_unique_phone",
-            phone,
-            unique=True,
-            postgresql_where=phone_verification_verified != None,
-        ),
-        Index(
-            "ix_users_active",
-            id,
-            postgresql_where=and_(not_(is_banned), not_(is_deleted)),
-        ),
-        # create index on users(geom, id, username) where not is_banned and not is_deleted and geom is not null;
-        Index(
-            "ix_users_geom_active",
-            geom,
-            id,
-            username,
-            postgresql_using="gist",
-            postgresql_where=and_(not_(is_banned), not_(is_deleted)),
-        ),
-        Index(
-            "ix_users_by_id",
-            id,
-            postgresql_using="hash",
-            postgresql_where=and_(not_(is_banned), not_(is_deleted)),
-        ),
-        Index(
-            "ix_users_by_username",
-            username,
-            postgresql_using="hash",
-            postgresql_where=and_(not_(is_banned), not_(is_deleted)),
-        ),
-        # There are two possible states for new_email_token, new_email_token_created, and new_email_token_expiry
-        CheckConstraint(
-            "(new_email_token IS NOT NULL AND new_email_token_created IS NOT NULL AND new_email_token_expiry IS NOT NULL) OR \
-             (new_email_token IS NULL AND new_email_token_created IS NULL AND new_email_token_expiry IS NULL)",
-            name="check_new_email_token_state",
-        ),
-        # Whenever a phone number is set, it must either be pending verification or already verified.
-        # Exactly one of the following must always be true: not phone, token, verified.
-        CheckConstraint(
-            "(phone IS NULL)::int + (phone_verification_verified IS NOT NULL)::int + (phone_verification_token IS NOT NULL)::int = 1",
-            name="phone_verified_conditions",
-        ),
-        # Email must match our regex
-        CheckConstraint(
-            f"email ~ '{EMAIL_REGEX}'",
-            name="valid_email",
-        ),
-        # Undelete token + time are coupled: either both null or neither; and if they're not null then the account is deleted
-        CheckConstraint(
-            "((undelete_token IS NULL) = (undelete_until IS NULL)) AND ((undelete_token IS NULL) OR is_deleted)",
-            name="undelete_nullity",
-        ),
-        # If the user disabled all emails, then they can't host or meet up
-        CheckConstraint(
-            "(do_not_email IS FALSE) OR ((hosting_status = 'cant_host') AND (meetup_status = 'does_not_want_to_meetup'))",
-            name="do_not_email_inactive",
-        ),
-    )
-
-    @hybrid_property
-    def has_completed_profile(self):
-        return self.avatar_key is not None and self.about_me is not None and len(self.about_me) >= 150
-
-    @has_completed_profile.expression
-    def has_completed_profile(cls):
-        return (cls.avatar_key != None) & (func.character_length(cls.about_me) >= 150)
-
-    @hybrid_property
-    def has_completed_my_home(self):
-        # completed my profile means that:
-        # 1. has filled out max_guests
-        # 2. has filled out sleeping_arrangement (sleeping privacy)
-        # 3. has some text in at least one of the my home free text fields
-        return (
-            self.max_guests is not None
-            and self.sleeping_arrangement is not None
-            and (
-                self.about_place is not None
-                or self.other_host_info is not None
-                or self.sleeping_details is not None
-                or self.area is not None
-                or self.house_rules is not None
-            )
-        )
-
-    @has_completed_my_home.expression
-    def has_completed_my_home(cls):
-        return and_(
-            cls.max_guests != None,
-            cls.sleeping_arrangement != None,
-            or_(
-                cls.about_place != None,
-                cls.other_host_info != None,
-                cls.sleeping_details != None,
-                cls.area != None,
-                cls.house_rules != None,
-            ),
-        )
-
-    @hybrid_property
-    def jailed_missing_tos(self):
-        return self.accepted_tos < TOS_VERSION
-
-    @hybrid_property
-    def jailed_missing_community_guidelines(self):
-        return self.accepted_community_guidelines < GUIDELINES_VERSION
-
-    @hybrid_property
-    def jailed_pending_mod_notes(self):
-        return self.mod_notes.where(ModNote.is_pending).count() > 0
-
-    @hybrid_property
-    def jailed_pending_activeness_probe(self):
-        return self.pending_activeness_probe != None
-
-    @hybrid_property
-    def is_jailed(self):
-        return (
-            self.jailed_missing_tos
-            | self.jailed_missing_community_guidelines
-            | self.is_missing_location
-            | self.jailed_pending_mod_notes
-            | self.jailed_pending_activeness_probe
-        )
-
-    @hybrid_property
-    def is_missing_location(self):
-        return self.needs_to_update_location
-
-    @hybrid_property
-    def is_visible(self):
-        return not self.is_banned and not self.is_deleted
-
-    @is_visible.expression
-    def is_visible(cls):
-        return ~(cls.is_banned | cls.is_deleted)
-
-    @property
-    def coordinates(self):
-        return get_coordinates(self.geom)
-
-    @property
-    def display_joined(self):
-        """
-        Returns the last active time rounded down to the nearest hour.
-        """
-        return self.joined.replace(minute=0, second=0, microsecond=0)
-
-    @property
-    def display_last_active(self):
-        """
-        Returns the last active time rounded down whatever is the "last active" coarsening.
-        """
-        return last_active_coarsen(self.last_active)
-
-    @hybrid_property
-    def phone_is_verified(self):
-        return (
-            self.phone_verification_verified is not None
-            and now() - self.phone_verification_verified < PHONE_VERIFICATION_LIFETIME
-        )
-
-    @phone_is_verified.expression
-    def phone_is_verified(cls):
-        return (cls.phone_verification_verified != None) & (
-            now() - cls.phone_verification_verified < PHONE_VERIFICATION_LIFETIME
-        )
-
-    @hybrid_property
-    def phone_code_expired(self):
-        return now() - self.phone_verification_sent > SMS_CODE_LIFETIME
-
-    def __repr__(self):
-        return f"User(id={self.id}, email={self.email}, username={self.username})"
-
-
-class UserBadge(Base):
-    """
-    A badge on a user's profile
-    """
-
-    __tablename__ = "user_badges"
-    __table_args__ = (UniqueConstraint("user_id", "badge_id"),)
-
-    id = Column(BigInteger, primary_key=True)
-
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    # corresponds to "id" in badges.json
-    badge_id = Column(String, nullable=False, index=True)
-
-    # take this with a grain of salt, someone may get then lose a badge for whatever reason
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    user = relationship("User", backref="badges")
-
-
-class ActivenessProbeStatus(enum.Enum):
-    # no response yet
-    pending = enum.auto()
-
-    # didn't respond on time
-    expired = enum.auto()
-
-    # responded that they're still active
-    still_active = enum.auto()
-
-    # responded that they're no longer active
-    no_longer_active = enum.auto()
-
-
-class ActivenessProbe(Base):
-    """
-    Activeness probes are used to gauge if users are still active: we send them a notification and ask them to respond,
-    we use this data both to help indicate response rate, as well as to make sure only those who are actively hosting
-    show up as such.
-    """
-
-    __tablename__ = "activeness_probes"
-
-    id = Column(BigInteger, primary_key=True)
-
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    # the time this probe was initiated
-    probe_initiated = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    # the number of reminders sent for this probe
-    notifications_sent = Column(Integer, nullable=False, server_default="0")
-
-    # the time of response
-    responded = Column(DateTime(timezone=True), nullable=True, default=None)
-    # the response value
-    response = Column(Enum(ActivenessProbeStatus), nullable=False, default=ActivenessProbeStatus.pending)
-
-    @hybrid_property
-    def is_pending(self):
-        return self.responded == None
-
-    user = relationship("User", back_populates="pending_activeness_probe")
-
-    __table_args__ = (
-        # a user can have at most one pending activeness probe at a time
-        Index(
-            "ix_activeness_probe_unique_pending_response",
-            user_id,
-            unique=True,
-            postgresql_where=responded == None,
-        ),
-        # response time is none iff response is pending
-        CheckConstraint(
-            "(responded IS NULL AND response = 'pending') OR (responded IS NOT NULL AND response != 'pending')",
-            name="pending_has_no_responded",
-        ),
-    )
-
-
-User.pending_activeness_probe = relationship(
-    ActivenessProbe,
-    primaryjoin="and_(ActivenessProbe.user_id == User.id, ActivenessProbe.is_pending)",
-    uselist=False,
-    back_populates="user",
-)
-
-
-class StrongVerificationAttemptStatus(enum.Enum):
-    ## full data states
-    # completed, this now provides verification for a user
-    succeeded = enum.auto()
-
-    ## no data states
-    # in progress: waiting for the user to scan the Iris code or open the app
-    in_progress_waiting_on_user_to_open_app = enum.auto()
-    # in progress: waiting for the user to scan MRZ or NFC/chip
-    in_progress_waiting_on_user_in_app = enum.auto()
-    # in progress, waiting for backend to pull verification data
-    in_progress_waiting_on_backend = enum.auto()
-    # failed, no data
-    failed = enum.auto()
-
-    # duplicate, at our end, has data
-    duplicate = enum.auto()
-
-    ## minimal data states
-    # the data, except minimal deduplication data, was deleted
-    deleted = enum.auto()
-
-
-class PassportSex(enum.Enum):
-    """
-    We don't care about sex, we use gender on the platform. But passports apparently do.
-    """
-
-    male = enum.auto()
-    female = enum.auto()
-    unspecified = enum.auto()
-
-
-class StrongVerificationAttempt(Base):
-    """
-    An attempt to perform strong verification
-    """
-
-    __tablename__ = "strong_verification_attempts"
-
-    # our verification id
-    id = Column(BigInteger, primary_key=True)
-
-    # this is returned in the callback, and we look up the attempt via this
-    verification_attempt_token = Column(String, nullable=False, unique=True)
-
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    status = Column(
-        Enum(StrongVerificationAttemptStatus),
-        nullable=False,
-        default=StrongVerificationAttemptStatus.in_progress_waiting_on_user_to_open_app,
-    )
-
-    ## full data
-    has_full_data = Column(Boolean, nullable=False, default=False)
-    # the data returned from iris, encrypted with a public key whose private key is kept offline
-    passport_encrypted_data = Column(Binary, nullable=True)
-    passport_date_of_birth = Column(Date, nullable=True)
-    passport_sex = Column(Enum(PassportSex), nullable=True)
-
-    ## minimal data: this will not be deleted
-    has_minimal_data = Column(Boolean, nullable=False, default=False)
-    passport_expiry_date = Column(Date, nullable=True)
-    passport_nationality = Column(String, nullable=True)
-    # last three characters of the passport number
-    passport_last_three_document_chars = Column(String, nullable=True)
-
-    iris_token = Column(String, nullable=False, unique=True)
-    iris_session_id = Column(BigInteger, nullable=False, unique=True)
-
-    passport_expiry_datetime = column_property(date_in_timezone(passport_expiry_date, "Etc/UTC"))
-
-    user = relationship("User")
-
-    @hybrid_property
-    def is_valid(self):
-        """
-        This only checks whether the attempt is a success and the passport is not expired, use `has_strong_verification` for full check
-        """
-        return (self.status == StrongVerificationAttemptStatus.succeeded) and (self.passport_expiry_datetime >= now())
-
-    @is_valid.expression
-    def is_valid(cls):
-        return (cls.status == StrongVerificationAttemptStatus.succeeded) & (
-            func.coalesce(cls.passport_expiry_datetime >= func.now(), False)
-        )
-
-    @hybrid_property
-    def is_visible(self):
-        return self.status != StrongVerificationAttemptStatus.deleted
-
-    @hybrid_method
-    def _raw_birthdate_match(self, user):
-        """Does not check whether the SV attempt itself is not expired"""
-        return self.passport_date_of_birth == user.birthdate
-
-    @hybrid_method
-    def matches_birthdate(self, user):
-        return self.is_valid & self._raw_birthdate_match(user)
-
-    @hybrid_method
-    def _raw_gender_match(self, user):
-        """Does not check whether the SV attempt itself is not expired"""
-        return (
-            ((user.gender == "Woman") & (self.passport_sex == PassportSex.female))
-            | ((user.gender == "Man") & (self.passport_sex == PassportSex.male))
-            | (self.passport_sex == PassportSex.unspecified)
-            | (user.has_passport_sex_gender_exception == True)
-        )
-
-    @hybrid_method
-    def matches_gender(self, user):
-        return self.is_valid & self._raw_gender_match(user)
-
-    @hybrid_method
-    def has_strong_verification(self, user):
-        return self.is_valid & self._raw_birthdate_match(user) & self._raw_gender_match(user)
-
-    __table_args__ = (
-        # used to look up verification status for a user
-        Index(
-            "ix_strong_verification_attempts_current",
-            user_id,
-            passport_expiry_date,
-            postgresql_where=status == StrongVerificationAttemptStatus.succeeded,
-        ),
-        # each passport can be verified only once
-        Index(
-            "ix_strong_verification_attempts_unique_succeeded",
-            passport_expiry_date,
-            passport_nationality,
-            passport_last_three_document_chars,
-            unique=True,
-            postgresql_where=(
-                (status == StrongVerificationAttemptStatus.succeeded)
-                | (status == StrongVerificationAttemptStatus.deleted)
-            ),
-        ),
-        # full data check
-        CheckConstraint(
-            "(has_full_data IS TRUE AND passport_encrypted_data IS NOT NULL AND passport_date_of_birth IS NOT NULL) OR \
-             (has_full_data IS FALSE AND passport_encrypted_data IS NULL AND passport_date_of_birth IS NULL)",
-            name="full_data_status",
-        ),
-        # minimal data check
-        CheckConstraint(
-            "(has_minimal_data IS TRUE AND passport_expiry_date IS NOT NULL AND passport_nationality IS NOT NULL AND passport_last_three_document_chars IS NOT NULL) OR \
-             (has_minimal_data IS FALSE AND passport_expiry_date IS NULL AND passport_nationality IS NULL AND passport_last_three_document_chars IS NULL)",
-            name="minimal_data_status",
-        ),
-        # note on implications: p => q iff ~p OR q
-        # full data implies minimal data, has_minimal_data => has_full_data
-        CheckConstraint(
-            "(has_full_data IS FALSE) OR (has_minimal_data IS TRUE)",
-            name="full_data_implies_minimal_data",
-        ),
-        # succeeded implies full data
-        CheckConstraint(
-            "(NOT (status = 'succeeded')) OR (has_full_data IS TRUE)",
-            name="succeeded_implies_full_data",
-        ),
-        # in_progress/failed implies no_data
-        CheckConstraint(
-            "(NOT ((status = 'in_progress_waiting_on_user_to_open_app') OR (status = 'in_progress_waiting_on_user_in_app') OR (status = 'in_progress_waiting_on_backend') OR (status = 'failed'))) OR (has_minimal_data IS FALSE)",
-            name="in_progress_failed_iris_implies_no_data",
-        ),
-        # deleted or duplicate implies minimal data
-        CheckConstraint(
-            "(NOT ((status = 'deleted') OR (status = 'duplicate'))) OR (has_minimal_data IS TRUE)",
-            name="deleted_duplicate_implies_minimal_data",
-        ),
-    )
-
-
-class ModNote(Base):
-    """
-    A moderator note to a user. This could be a warning, just a note "hey, we did X", or any other similar message.
-
-    The user has to read and "acknowledge" the note before continuing onto the platform, and being un-jailed.
-    """
-
-    __tablename__ = "mod_notes"
-    id = Column(BigInteger, primary_key=True)
-
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    acknowledged = Column(DateTime(timezone=True), nullable=True)
-
-    # this is an internal ID to allow the mods to track different types of notes
-    internal_id = Column(String, nullable=False)
-    # the admin that left this note
-    creator_user_id = Column(ForeignKey("users.id"), nullable=False)
-
-    note_content = Column(String, nullable=False)  # CommonMark without images
-
-    user = relationship("User", backref=backref("mod_notes", lazy="dynamic"), foreign_keys="ModNote.user_id")
-
-    def __repr__(self):
-        return f"ModeNote(id={self.id}, user={self.user}, created={self.created}, ack'd={self.acknowledged})"
-
-    @hybrid_property
-    def is_pending(self):
-        return self.acknowledged == None
-
-    __table_args__ = (
-        # used to look up pending notes
-        Index(
-            "ix_mod_notes_unacknowledged",
-            user_id,
-            postgresql_where=acknowledged == None,
-        ),
-    )
-
-
-class StrongVerificationCallbackEvent(Base):
-    __tablename__ = "strong_verification_callback_events"
-
-    id = Column(BigInteger, primary_key=True)
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    verification_attempt_id = Column(ForeignKey("strong_verification_attempts.id"), nullable=False, index=True)
-
-    iris_status = Column(String, nullable=False)
-
-
-class DonationType(enum.Enum):
-    one_time = enum.auto()
-    recurring = enum.auto()
-
-
-class DonationInitiation(Base):
-    """
-    Whenever someone initiaties a donation through the platform
-    """
-
-    __tablename__ = "donation_initiations"
-    id = Column(BigInteger, primary_key=True)
-
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-
-    amount = Column(Integer, nullable=False)
-    stripe_checkout_session_id = Column(String, nullable=False)
-
-    donation_type = Column(Enum(DonationType), nullable=False)
-    source = Column(String, nullable=True)
-
-    user = relationship("User", backref="donation_initiations")
-
-
-class Invoice(Base):
-    """
-    Successful donations, both one off and recurring
-
-    Triggered by `payment_intent.succeeded` webhook
-    """
-
-    __tablename__ = "invoices"
-
-    id = Column(BigInteger, primary_key=True)
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    user_id = Column(ForeignKey("users.id"), nullable=False)
-
-    amount = Column(Float, nullable=False)
-
-    stripe_payment_intent_id = Column(String, nullable=False, unique=True)
-    stripe_receipt_url = Column(String, nullable=False)
-
-    user = relationship("User", backref="invoices")
-
-
-class LanguageFluency(enum.Enum):
-    # note that the numbering is important here, these are ordinal
-    beginner = 1
-    conversational = 2
-    fluent = 3
-
-
-class LanguageAbility(Base):
-    __tablename__ = "language_abilities"
-    __table_args__ = (
-        # Users can only have one language ability per language
-        UniqueConstraint("user_id", "language_code"),
-    )
-
-    id = Column(BigInteger, primary_key=True)
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    language_code = Column(ForeignKey("languages.code", deferrable=True), nullable=False)
-    fluency = Column(Enum(LanguageFluency), nullable=False)
-
-    user = relationship("User", backref="language_abilities")
-    language = relationship("Language")
-
-
-class RegionVisited(Base):
-    __tablename__ = "regions_visited"
-    __table_args__ = (UniqueConstraint("user_id", "region_code"),)
-
-    id = Column(BigInteger, primary_key=True)
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    region_code = Column(ForeignKey("regions.code", deferrable=True), nullable=False)
-
-
-class RegionLived(Base):
-    __tablename__ = "regions_lived"
-    __table_args__ = (UniqueConstraint("user_id", "region_code"),)
-
-    id = Column(BigInteger, primary_key=True)
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    region_code = Column(ForeignKey("regions.code", deferrable=True), nullable=False)
-
-
-class FriendStatus(enum.Enum):
-    pending = enum.auto()
-    accepted = enum.auto()
-    rejected = enum.auto()
-    cancelled = enum.auto()
-
-
-class FriendRelationship(Base):
-    """
-    Friendship relations between users
-
-    TODO: make this better with sqlalchemy self-referential stuff
-    TODO: constraint on only one row per user pair where accepted or pending
-    """
-
-    __tablename__ = "friend_relationships"
-
-    id = Column(BigInteger, primary_key=True)
-
-    from_user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    to_user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-
-    status = Column(Enum(FriendStatus), nullable=False, default=FriendStatus.pending)
-
-    # timezones should always be UTC
-    time_sent = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    time_responded = Column(DateTime(timezone=True), nullable=True)
-
-    from_user = relationship("User", backref="friends_from", foreign_keys="FriendRelationship.from_user_id")
-    to_user = relationship("User", backref="friends_to", foreign_keys="FriendRelationship.to_user_id")
-
-    __table_args__ = (
-        # Ping looks up pending friend reqs, this speeds that up
-        Index(
-            "ix_friend_relationships_status_to_from",
-            status,
-            to_user_id,
-            from_user_id,
-        ),
-    )
-
-
-class ContributeOption(enum.Enum):
-    yes = enum.auto()
-    maybe = enum.auto()
-    no = enum.auto()
-
-
-class ContributorForm(Base):
-    """
-    Someone filled in the contributor form
-    """
-
-    __tablename__ = "contributor_forms"
-
-    id = Column(BigInteger, primary_key=True)
-
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    ideas = Column(String, nullable=True)
-    features = Column(String, nullable=True)
-    experience = Column(String, nullable=True)
-    contribute = Column(Enum(ContributeOption), nullable=True)
-    contribute_ways = Column(ARRAY(String), nullable=False)
-    expertise = Column(String, nullable=True)
-
-    user = relationship("User", backref="contributor_forms")
-
-    @hybrid_property
-    def is_filled(self):
-        """
-        Whether the form counts as having been filled
-        """
-        return (
-            (self.ideas != None)
-            | (self.features != None)
-            | (self.experience != None)
-            | (self.contribute != None)
-            | (self.contribute_ways != [])
-            | (self.expertise != None)
-        )
-
-    @property
-    def should_notify(self):
-        """
-        If this evaluates to true, we send an email to the recruitment team.
-
-        We currently send if expertise is listed, or if they list a way to help outside of a set list
-        """
-        return False
-
-
-class SignupFlow(Base):
-    """
-    Signup flows/incomplete users
-
-    Coinciding fields have the same meaning as in User
-    """
-
-    __tablename__ = "signup_flows"
-
-    id = Column(BigInteger, primary_key=True)
-
-    # housekeeping
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    flow_token = Column(String, nullable=False, unique=True)
-    email_verified = Column(Boolean, nullable=False, default=False)
-    email_sent = Column(Boolean, nullable=False, default=False)
-    email_token = Column(String, nullable=True)
-    email_token_expiry = Column(DateTime(timezone=True), nullable=True)
-
-    ## Basic
-    name = Column(String, nullable=False)
-    # TODO: unique across both tables
-    email = Column(String, nullable=False, unique=True)
-    # TODO: invitation, attribution
-
-    ## Account
-    # TODO: unique across both tables
-    username = Column(String, nullable=True, unique=True)
-    hashed_password = Column(Binary, nullable=True)
-    birthdate = Column(Date, nullable=True)  # in the timezone of birthplace
-    gender = Column(String, nullable=True)
-    hosting_status = Column(Enum(HostingStatus), nullable=True)
-    city = Column(String, nullable=True)
-    geom = Column(Geometry(geometry_type="POINT", srid=4326), nullable=True)
-    geom_radius = Column(Float, nullable=True)
-
-    accepted_tos = Column(Integer, nullable=True)
-    accepted_community_guidelines = Column(Integer, nullable=False, server_default="0")
-
-    opt_out_of_newsletter = Column(Boolean, nullable=True)
-
-    ## Feedback (now unused)
-    filled_feedback = Column(Boolean, nullable=False, default=False)
-    ideas = Column(String, nullable=True)
-    features = Column(String, nullable=True)
-    experience = Column(String, nullable=True)
-    contribute = Column(Enum(ContributeOption), nullable=True)
-    contribute_ways = Column(ARRAY(String), nullable=True)
-    expertise = Column(String, nullable=True)
-
-    invite_code_id = Column(ForeignKey("invite_codes.id"), nullable=True)
-
-    @hybrid_property
-    def token_is_valid(self):
-        return (self.email_token != None) & (self.email_token_expiry >= now())
-
-    @hybrid_property
-    def account_is_filled(self):
-        return (
-            (self.username != None)
-            & (self.birthdate != None)
-            & (self.gender != None)
-            & (self.hosting_status != None)
-            & (self.city != None)
-            & (self.geom != None)
-            & (self.geom_radius != None)
-            & (self.accepted_tos != None)
-            & (self.opt_out_of_newsletter != None)
-        )
-
-    @hybrid_property
-    def is_completed(self):
-        return self.email_verified & self.account_is_filled & (self.accepted_community_guidelines == GUIDELINES_VERSION)
-
-
-class LoginToken(Base):
-    """
-    A login token sent in an email to a user, allows them to sign in between the times defined by created and expiry
-    """
-
-    __tablename__ = "login_tokens"
-    token = Column(String, primary_key=True)
-
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-
-    # timezones should always be UTC
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    expiry = Column(DateTime(timezone=True), nullable=False)
-
-    user = relationship("User", backref="login_tokens")
-
-    @hybrid_property
-    def is_valid(self):
-        return (self.created <= now()) & (self.expiry >= now())
-
-    def __repr__(self):
-        return f"LoginToken(token={self.token}, user={self.user}, created={self.created}, expiry={self.expiry})"
-
-
-class PasswordResetToken(Base):
-    __tablename__ = "password_reset_tokens"
-    token = Column(String, primary_key=True)
-
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    expiry = Column(DateTime(timezone=True), nullable=False)
-
-    user = relationship("User", backref="password_reset_tokens")
-
-    @hybrid_property
-    def is_valid(self):
-        return (self.created <= now()) & (self.expiry >= now())
-
-    def __repr__(self):
-        return f"PasswordResetToken(token={self.token}, user={self.user}, created={self.created}, expiry={self.expiry})"
-
-
-class AccountDeletionToken(Base):
-    __tablename__ = "account_deletion_tokens"
-
-    token = Column(String, primary_key=True)
-
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    expiry = Column(DateTime(timezone=True), nullable=False)
-
-    user = relationship("User", backref="account_deletion_tokens")
-
-    @hybrid_property
-    def is_valid(self):
-        return (self.created <= now()) & (self.expiry >= now())
-
-    def __repr__(self):
-        return f"AccountDeletionToken(token={self.token}, user_id={self.user_id}, created={self.created}, expiry={self.expiry})"
-
-
-class UserActivity(Base):
-    """
-    User activity: for each unique (user_id, period, ip_address, user_agent) tuple, keep track of number of api calls
-
-    Used for user "last active" as well as admin stuff
-    """
-
-    __tablename__ = "user_activity"
-
-    id = Column(BigInteger, primary_key=True)
-
-    user_id = Column(ForeignKey("users.id"), nullable=False)
-    # the start of a period of time, e.g. 1 hour during which we bin activeness
-    period = Column(DateTime(timezone=True), nullable=False)
-
-    # details of the browser, if available
-    ip_address = Column(INET, nullable=True)
-    user_agent = Column(String, nullable=True)
-
-    # count of api calls made with this ip, user_agent, and period
-    api_calls = Column(Integer, nullable=False, default=0)
-
-    __table_args__ = (
-        # helps look up this tuple quickly
-        Index(
-            "ix_user_activity_user_id_period_ip_address_user_agent",
-            user_id,
-            period,
-            ip_address,
-            user_agent,
-            unique=True,
-        ),
-    )
-
-
-class UserSession(Base):
-    """
-    API keys/session cookies for the app
-
-    There are two types of sessions: long-lived, and short-lived. Long-lived are
-    like when you choose "remember this browser": they will be valid for a long
-    time without the user interacting with the site. Short-lived sessions on the
-    other hand get invalidated quickly if the user does not interact with the
-    site.
-
-    Long-lived tokens are valid from `created` until `expiry`.
-
-    Short-lived tokens expire after 168 hours (7 days) if they are not used.
-    """
-
-    __tablename__ = "sessions"
-    token = Column(String, primary_key=True)
-
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-
-    # sessions are either "api keys" or "session cookies", otherwise identical
-    # an api key is put in the authorization header (e.g. "authorization: Bearer <token>")
-    # a session cookie is set in the "couchers-sesh" cookie (e.g. "cookie: couchers-sesh=<token>")
-    # when a session is created, it's fixed as one or the other for security reasons
-    # for api keys to be useful, they should be long lived and have a long expiry
-    is_api_key = Column(Boolean, nullable=False, server_default=expression.false())
-
-    # whether it's a long-lived or short-lived session
-    long_lived = Column(Boolean, nullable=False)
-
-    # the time at which the session was created
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    # the expiry of the session: the session *cannot* be refreshed past this
-    expiry = Column(DateTime(timezone=True), nullable=False, server_default=func.now() + text("interval '730 days'"))
-
-    # the time at which the token was invalidated, allows users to delete sessions
-    deleted = Column(DateTime(timezone=True), nullable=True, default=None)
-
-    # the last time this session was used
-    last_seen = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    # count of api calls made with this token/session (if we're updating last_seen, might as well update this too)
-    api_calls = Column(Integer, nullable=False, default=0)
-
-    # details of the browser, if available
-    # these are from the request creating the session, not used for anything else
-    ip_address = Column(String, nullable=True)
-    user_agent = Column(String, nullable=True)
-
-    user = relationship("User", backref="sessions")
-
-    @hybrid_property
-    def is_valid(self):
-        """
-        It must have been created and not be expired or deleted.
-
-        Also, if it's a short lived token, it must have been used in the last 168 hours.
-
-        TODO: this probably won't run in python (instance level), only in sql (class level)
-        """
-        return (
-            (self.created <= func.now())
-            & (self.expiry >= func.now())
-            & (self.deleted == None)
-            & (self.long_lived | (func.now() - self.last_seen < text("interval '168 hours'")))
-        )
-
-    __table_args__ = (
-        Index(
-            "ix_sessions_by_token",
-            "token",
-            postgresql_using="hash",
-        ),
-    )
-
-
-class Conversation(Base):
-    """
-    Conversation brings together the different types of message/conversation types
-    """
-
-    __tablename__ = "conversations"
-
-    id = Column(BigInteger, primary_key=True)
-    # timezone should always be UTC
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    def __repr__(self):
-        return f"Conversation(id={self.id}, created={self.created})"
-
-
-class GroupChat(Base):
-    """
-    Group chat
-    """
-
-    __tablename__ = "group_chats"
-
-    conversation_id = Column("id", ForeignKey("conversations.id"), nullable=False, primary_key=True)
-
-    title = Column(String, nullable=True)
-    only_admins_invite = Column(Boolean, nullable=False, default=True)
-    creator_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    is_dm = Column(Boolean, nullable=False)
-
-    conversation = relationship("Conversation", backref="group_chat")
-    creator = relationship("User", backref="created_group_chats")
-
-    def __repr__(self):
-        return f"GroupChat(conversation={self.conversation}, title={self.title or 'None'}, only_admins_invite={self.only_admins_invite}, creator={self.creator}, is_dm={self.is_dm})"
-
-
-class GroupChatRole(enum.Enum):
-    admin = enum.auto()
-    participant = enum.auto()
-
-
-class GroupChatSubscription(Base):
-    """
-    The recipient of a thread and information about when they joined/left/etc.
-    """
-
-    __tablename__ = "group_chat_subscriptions"
-    id = Column(BigInteger, primary_key=True)
-
-    # TODO: DB constraint on only one user+group_chat combo at a given time
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    group_chat_id = Column(ForeignKey("group_chats.id"), nullable=False, index=True)
-
-    # timezones should always be UTC
-    joined = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    left = Column(DateTime(timezone=True), nullable=True)
-
-    role = Column(Enum(GroupChatRole), nullable=False)
-
-    last_seen_message_id = Column(BigInteger, nullable=False, default=0)
-
-    # when this chat is muted until, DATETIME_INFINITY for "forever"
-    muted_until = Column(DateTime(timezone=True), nullable=False, server_default=DATETIME_MINUS_INFINITY.isoformat())
-
-    user = relationship("User", backref="group_chat_subscriptions")
-    group_chat = relationship("GroupChat", backref=backref("subscriptions", lazy="dynamic"))
-
-    def muted_display(self):
-        """
-        Returns (muted, muted_until) display values:
-        1. If not muted, returns (False, None)
-        2. If muted forever, returns (True, None)
-        3. If muted until a given datetime returns (True, dt)
-        """
-        if self.muted_until < now():
-            return (False, None)
-        elif self.muted_until == DATETIME_INFINITY:
-            return (True, None)
-        else:
-            return (True, self.muted_until)
-
-    @hybrid_property
-    def is_muted(self):
-        return self.muted_until > func.now()
-
-    def __repr__(self):
-        return f"GroupChatSubscription(id={self.id}, user={self.user}, joined={self.joined}, left={self.left}, role={self.role}, group_chat={self.group_chat})"
-
-
-class InviteCode(Base):
-    __tablename__ = "invite_codes"
-
-    id = Column(String, primary_key=True)
-    creator_user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
-    created = Column(DateTime(timezone=True), nullable=False, default=func.now())
-    disabled = Column(DateTime(timezone=True), nullable=True)
-
-    creator = relationship("User", foreign_keys=[creator_user_id])
-
-
-class MessageType(enum.Enum):
-    text = enum.auto()
-    # e.g.
-    # image =
-    # emoji =
-    # ...
-    chat_created = enum.auto()
-    chat_edited = enum.auto()
-    user_invited = enum.auto()
-    user_left = enum.auto()
-    user_made_admin = enum.auto()
-    user_removed_admin = enum.auto()  # RemoveGroupChatAdmin: remove admin permission from a user in group chat
-    host_request_status_changed = enum.auto()
-    user_removed = enum.auto()  # user is removed from group chat by amdin RemoveGroupChatUser
-
-
-class HostRequestStatus(enum.Enum):
-    pending = enum.auto()
-    accepted = enum.auto()
-    rejected = enum.auto()
-    confirmed = enum.auto()
-    cancelled = enum.auto()
-
-
-class Message(Base):
-    """
-    A message.
-
-    If message_type = text, then the message is a normal text message, otherwise, it's a special control message.
-    """
-
-    __tablename__ = "messages"
-
-    id = Column(BigInteger, primary_key=True)
-
-    # which conversation the message belongs in
-    conversation_id = Column(ForeignKey("conversations.id"), nullable=False, index=True)
-
-    # the user that sent the message/command
-    author_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-
-    # the message type, "text" is a text message, otherwise a "control message"
-    message_type = Column(Enum(MessageType), nullable=False)
-
-    # the target if a control message and requires target, e.g. if inviting a user, the user invited is the target
-    target_id = Column(ForeignKey("users.id"), nullable=True, index=True)
-
-    # time sent, timezone should always be UTC
-    time = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    # the plain-text message text if not control
-    text = Column(String, nullable=True)
-
-    # the new host request status if the message type is host_request_status_changed
-    host_request_status_target = Column(Enum(HostRequestStatus), nullable=True)
-
-    conversation = relationship("Conversation", backref="messages", order_by="Message.time.desc()")
-    author = relationship("User", foreign_keys="Message.author_id")
-    target = relationship("User", foreign_keys="Message.target_id")
-
-    @property
-    def is_normal_message(self):
-        """
-        There's only one normal type atm, text
-        """
-        return self.message_type == MessageType.text
-
-    def __repr__(self):
-        return f"Message(id={self.id}, time={self.time}, text={self.text}, author={self.author}, conversation={self.conversation})"
-
-
-class ContentReport(Base):
-    """
-    A piece of content reported to admins
-    """
-
-    __tablename__ = "content_reports"
-
-    id = Column(BigInteger, primary_key=True)
-
-    time = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    # the user who reported or flagged the content
-    reporting_user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-
-    # reason, e.g. spam, inappropriate, etc
-    reason = Column(String, nullable=False)
-    # a short description
-    description = Column(String, nullable=False)
-
-    # a reference to the content, see //docs/content_ref.md
-    content_ref = Column(String, nullable=False)
-    # the author of the content (e.g. the user who wrote the comment itself)
-    author_user_id = Column(ForeignKey("users.id"), nullable=False)
-
-    # details of the browser, if available
-    user_agent = Column(String, nullable=False)
-    # the URL the user was on when reporting the content
-    page = Column(String, nullable=False)
-
-    # see comments above for reporting vs author
-    reporting_user = relationship("User", foreign_keys="ContentReport.reporting_user_id")
-    author_user = relationship("User", foreign_keys="ContentReport.author_user_id")
-
-
-class Email(Base):
-    """
-    Table of all dispatched emails for debugging purposes, etc.
-    """
-
-    __tablename__ = "emails"
-
-    id = Column(String, primary_key=True)
-
-    # timezone should always be UTC
-    time = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    sender_name = Column(String, nullable=False)
-    sender_email = Column(String, nullable=False)
-
-    recipient = Column(String, nullable=False)
-    subject = Column(String, nullable=False)
-
-    plain = Column(String, nullable=False)
-    html = Column(String, nullable=False)
-
-    list_unsubscribe_header = Column(String, nullable=True)
-    source_data = Column(String, nullable=True)
-
-
-class SMS(Base):
-    """
-    Table of all sent SMSs for debugging purposes, etc.
-    """
-
-    __tablename__ = "smss"
-
-    id = Column(BigInteger, primary_key=True)
-
-    # timezone should always be UTC
-    time = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    # AWS message id
-    message_id = Column(String, nullable=False)
-
-    # the SMS sender ID sent to AWS, name that the SMS appears to come from
-    sms_sender_id = Column(String, nullable=False)
-    number = Column(String, nullable=False)
-    message = Column(String, nullable=False)
-
-
-class HostRequest(Base):
-    """
-    A request to stay with a host
-    """
-
-    __tablename__ = "host_requests"
-
-    conversation_id = Column("id", ForeignKey("conversations.id"), nullable=False, primary_key=True)
-    surfer_user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    host_user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-
-    hosting_city = Column(String, nullable=False)
-    hosting_location = Column(Geometry("POINT", srid=4326), nullable=False)
-    hosting_radius = Column(Float, nullable=False)
-
-    # TODO: proper timezone handling
-    timezone = "Etc/UTC"
-
-    # dates in the timezone above
-    from_date = Column(Date, nullable=False)
-    to_date = Column(Date, nullable=False)
-
-    # timezone aware start and end times of the request, can be compared to now()
-    start_time = column_property(date_in_timezone(from_date, timezone))
-    end_time = column_property(date_in_timezone(to_date, timezone) + text("interval '1 days'"))
-    start_time_to_write_reference = column_property(date_in_timezone(to_date, timezone))
-    # notice 1 day for midnight at the *end of the day*, then 14 days to write a ref
-    end_time_to_write_reference = column_property(date_in_timezone(to_date, timezone) + text("interval '15 days'"))
-
-    status = Column(Enum(HostRequestStatus), nullable=False)
-    is_host_archived = Column(Boolean, nullable=False, default=False, server_default=expression.false())
-    is_surfer_archived = Column(Boolean, nullable=False, default=False, server_default=expression.false())
-
-    host_last_seen_message_id = Column(BigInteger, nullable=False, default=0)
-    surfer_last_seen_message_id = Column(BigInteger, nullable=False, default=0)
-
-    # number of reference reminders sent out
-    host_sent_reference_reminders = Column(BigInteger, nullable=False, server_default=text("0"))
-    surfer_sent_reference_reminders = Column(BigInteger, nullable=False, server_default=text("0"))
-    host_sent_request_reminders = Column(BigInteger, nullable=False, server_default=text("0"))
-    last_sent_request_reminder_time = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    # reason why the host/surfer marked that they didn't meet up
-    # if null then they haven't marked it such
-    host_reason_didnt_meetup = Column(String, nullable=True)
-    surfer_reason_didnt_meetup = Column(String, nullable=True)
-
-    surfer = relationship("User", backref="host_requests_sent", foreign_keys="HostRequest.surfer_user_id")
-    host = relationship("User", backref="host_requests_received", foreign_keys="HostRequest.host_user_id")
-    conversation = relationship("Conversation")
-
-    __table_args__ = (
-        # allows fast lookup as to whether they didn't meet up
-        Index(
-            "ix_host_requests_host_didnt_meetup",
-            host_reason_didnt_meetup != None,
-        ),
-        Index(
-            "ix_host_requests_surfer_didnt_meetup",
-            surfer_reason_didnt_meetup != None,
-        ),
-        # Used for figuring out who needs a reminder to respond
-        Index(
-            "ix_host_requests_status_reminder_counts",
-            status,
-            host_sent_request_reminders,
-            last_sent_request_reminder_time,
-            from_date,
-        ),
-    )
-
-    @hybrid_property
-    def can_write_reference(self):
-        return (
-            ((self.status == HostRequestStatus.confirmed) | (self.status == HostRequestStatus.accepted))
-            & (now() >= self.start_time_to_write_reference)
-            & (now() <= self.end_time_to_write_reference)
-        )
-
-    @can_write_reference.expression
-    def can_write_reference(cls):
-        return (
-            ((cls.status == HostRequestStatus.confirmed) | (cls.status == HostRequestStatus.accepted))
-            & (func.now() >= cls.start_time_to_write_reference)
-            & (func.now() <= cls.end_time_to_write_reference)
-        )
-
-    def __repr__(self):
-        return f"HostRequest(id={self.conversation_id}, surfer_user_id={self.surfer_user_id}, host_user_id={self.host_user_id}...)"
-
-
-class HostRequestQuality(enum.Enum):
-    high_quality = enum.auto()
-    okay_quality = enum.auto()
-    low_quality = enum.auto()
-
-
-class HostRequestFeedback(Base):
-    """
-    Private feedback from host about a host request
-    """
-
-    __tablename__ = "host_request_feedbacks"
-
-    id = Column(BigInteger, primary_key=True)
-    time = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    host_request_id = Column(ForeignKey("host_requests.id"), nullable=False)
-
-    from_user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    to_user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-
-    request_quality = Column(Enum(HostRequestQuality), nullable=True)
-    decline_reason = Column(String, nullable=True)  # plain text
-
-    host_request = relationship("HostRequest")
-
-    __table_args__ = (
-        # Each user can leave at most one friend reference to another user
-        Index(
-            "ix_unique_host_req_feedback",
-            from_user_id,
-            to_user_id,
-            host_request_id,
-            unique=True,
-        ),
-    )
-
-
-class ReferenceType(enum.Enum):
-    friend = enum.auto()
-    surfed = enum.auto()  # The "from" user surfed with the "to" user
-    hosted = enum.auto()  # The "from" user hosted the "to" user
-
-
-class Reference(Base):
-    """
-    Reference from one user to another
-    """
-
-    __tablename__ = "references"
-
-    id = Column(BigInteger, primary_key=True)
-    # timezone should always be UTC
-    time = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    from_user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    to_user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-
-    reference_type = Column(Enum(ReferenceType), nullable=False)
-
-    host_request_id = Column(ForeignKey("host_requests.id"), nullable=True)
-
-    text = Column(String, nullable=False)  # plain text
-    # text that's only visible to mods
-    private_text = Column(String, nullable=True)  # plain text
-
-    rating = Column(Float, nullable=False)
-    was_appropriate = Column(Boolean, nullable=False)
-
-    is_deleted = Column(Boolean, nullable=False, default=False, server_default=expression.false())
-
-    from_user = relationship("User", backref="references_from", foreign_keys="Reference.from_user_id")
-    to_user = relationship("User", backref="references_to", foreign_keys="Reference.to_user_id")
-
-    host_request = relationship("HostRequest", backref="references")
-
-    __table_args__ = (
-        # Rating must be between 0 and 1, inclusive
-        CheckConstraint(
-            "rating BETWEEN 0 AND 1",
-            name="rating_between_0_and_1",
-        ),
-        # Has host_request_id or it's a friend reference
-        CheckConstraint(
-            "(host_request_id IS NOT NULL) <> (reference_type = 'friend')",
-            name="host_request_id_xor_friend_reference",
-        ),
-        # Each user can leave at most one friend reference to another user
-        Index(
-            "ix_references_unique_friend_reference",
-            from_user_id,
-            to_user_id,
-            reference_type,
-            unique=True,
-            postgresql_where=(reference_type == ReferenceType.friend),
-        ),
-        # Each user can leave at most one reference to another user for each stay
-        Index(
-            "ix_references_unique_per_host_request",
-            from_user_id,
-            to_user_id,
-            host_request_id,
-            unique=True,
-            postgresql_where=(host_request_id != None),
-        ),
-    )
-
-    @property
-    def should_report(self):
-        """
-        If this evaluates to true, we send a report to the moderation team.
-        """
-        return self.rating <= 0.4 or not self.was_appropriate or self.private_text
-
-
-class InitiatedUpload(Base):
-    """
-    Started downloads, not necessarily complete yet.
-    """
-
-    __tablename__ = "initiated_uploads"
-
-    key = Column(String, primary_key=True)
-
-    # timezones should always be UTC
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    expiry = Column(DateTime(timezone=True), nullable=False)
-
-    initiator_user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-
-    initiator_user = relationship("User")
-
-    @hybrid_property
-    def is_valid(self):
-        return (self.created <= func.now()) & (self.expiry >= func.now())
-
-
-class Upload(Base):
-    """
-    Completed uploads.
-    """
-
-    __tablename__ = "uploads"
-    key = Column(String, primary_key=True)
-
-    filename = Column(String, nullable=False)
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    creator_user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-
-    # photo credit, etc
-    credit = Column(String, nullable=True)
-
-    creator_user = relationship("User", backref="uploads", foreign_keys="Upload.creator_user_id")
-
-    def _url(self, size):
-        return urls.media_url(filename=self.filename, size=size)
-
-    @property
-    def thumbnail_url(self):
-        return self._url("thumbnail")
-
-    @property
-    def full_url(self):
-        return self._url("full")
-
-
-communities_seq = Sequence("communities_seq")
-
-
-class Node(Base):
-    """
-    Node, i.e. geographical subdivision of the world
-
-    Administered by the official cluster
-    """
-
-    __tablename__ = "nodes"
-
-    id = Column(BigInteger, communities_seq, primary_key=True, server_default=communities_seq.next_value())
-
-    # name and description come from official cluster
-    parent_node_id = Column(ForeignKey("nodes.id"), nullable=True, index=True)
-    geom = deferred(Column(Geometry(geometry_type="MULTIPOLYGON", srid=4326), nullable=False))
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    parent_node = relationship("Node", backref="child_nodes", remote_side="Node.id")
-
-    contained_users = relationship(
-        "User",
-        primaryjoin="func.ST_Contains(foreign(Node.geom), User.geom).as_comparison(1, 2)",
-        viewonly=True,
-        uselist=True,
-    )
-
-    contained_user_ids = association_proxy("contained_users", "id")
-
-
-class Cluster(Base):
-    """
-    Cluster, administered grouping of content
-    """
-
-    __tablename__ = "clusters"
-
-    id = Column(BigInteger, communities_seq, primary_key=True, server_default=communities_seq.next_value())
-    parent_node_id = Column(ForeignKey("nodes.id"), nullable=False, index=True)
-    name = Column(String, nullable=False)
-    # short description
-    description = Column(String, nullable=False)
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    is_official_cluster = Column(Boolean, nullable=False, default=False)
-
-    discussions_enabled = Column(Boolean, nullable=False, default=True, server_default=expression.true())
-    events_enabled = Column(Boolean, nullable=False, default=True, server_default=expression.true())
-
-    slug = column_property(func.slugify(name))
-
-    official_cluster_for_node = relationship(
-        "Node",
-        primaryjoin="and_(Cluster.parent_node_id == Node.id, Cluster.is_official_cluster)",
-        backref=backref("official_cluster", uselist=False),
-        uselist=False,
-        viewonly=True,
-    )
-
-    parent_node = relationship(
-        "Node", backref="child_clusters", remote_side="Node.id", foreign_keys="Cluster.parent_node_id"
-    )
-
-    nodes = relationship("Cluster", backref="clusters", secondary="node_cluster_associations", viewonly=True)
-    # all pages
-    pages = relationship(
-        "Page", backref="clusters", secondary="cluster_page_associations", lazy="dynamic", viewonly=True
-    )
-    events = relationship("Event", backref="clusters", secondary="cluster_event_associations", viewonly=True)
-    discussions = relationship(
-        "Discussion", backref="clusters", secondary="cluster_discussion_associations", viewonly=True
-    )
-
-    # includes also admins
-    members = relationship(
-        "User",
-        lazy="dynamic",
-        backref="cluster_memberships",
-        secondary="cluster_subscriptions",
-        primaryjoin="Cluster.id == ClusterSubscription.cluster_id",
-        secondaryjoin="User.id == ClusterSubscription.user_id",
-        viewonly=True,
-    )
-
-    admins = relationship(
-        "User",
-        lazy="dynamic",
-        backref="cluster_adminships",
-        secondary="cluster_subscriptions",
-        primaryjoin="Cluster.id == ClusterSubscription.cluster_id",
-        secondaryjoin="and_(User.id == ClusterSubscription.user_id, ClusterSubscription.role == 'admin')",
-        viewonly=True,
-    )
-
-    main_page = relationship(
-        "Page",
-        primaryjoin="and_(Cluster.id == Page.owner_cluster_id, Page.type == 'main_page')",
-        viewonly=True,
-        uselist=False,
-    )
-
-    @property
-    def is_leaf(self) -> bool:
-        """Whether the cluster is a leaf node in the cluster hierarchy."""
-        return len(self.parent_node.child_nodes) == 0
-
-    __table_args__ = (
-        # Each node can have at most one official cluster
-        Index(
-            "ix_clusters_owner_parent_node_id_is_official_cluster",
-            parent_node_id,
-            is_official_cluster,
-            unique=True,
-            postgresql_where=is_official_cluster,
-        ),
-        # trigram index on unaccented name
-        # note that the function `unaccent` is not immutable so cannot be used in an index, that's why we wrap it
-        Index(
-            "idx_clusters_name_unaccented_trgm",
-            text("immutable_unaccent(name) gin_trgm_ops"),
-            postgresql_using="gin",
-        ),
-    )
-
-
-class NodeClusterAssociation(Base):
-    """
-    NodeClusterAssociation, grouping of nodes
-    """
-
-    __tablename__ = "node_cluster_associations"
-    __table_args__ = (UniqueConstraint("node_id", "cluster_id"),)
-
-    id = Column(BigInteger, primary_key=True)
-
-    node_id = Column(ForeignKey("nodes.id"), nullable=False, index=True)
-    cluster_id = Column(ForeignKey("clusters.id"), nullable=False, index=True)
-
-    node = relationship("Node", backref="node_cluster_associations")
-    cluster = relationship("Cluster", backref="node_cluster_associations")
-
-
-class ClusterRole(enum.Enum):
-    member = enum.auto()
-    admin = enum.auto()
-
-
-class ClusterSubscription(Base):
-    """
-    ClusterSubscription of a user
-    """
-
-    __tablename__ = "cluster_subscriptions"
-
-    id = Column(BigInteger, primary_key=True)
-
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    cluster_id = Column(ForeignKey("clusters.id"), nullable=False, index=True)
-    role = Column(Enum(ClusterRole), nullable=False)
-
-    user = relationship("User", backref="cluster_subscriptions")
-    cluster = relationship("Cluster", backref="cluster_subscriptions")
-
-    __table_args__ = (
-        UniqueConstraint("user_id", "cluster_id"),
-        Index(
-            "ix_cluster_subscriptions_members",
-            cluster_id,
-            user_id,
-        ),
-        # For fast lookup of nodes this user is an admin of
-        Index(
-            "ix_cluster_subscriptions_admins",
-            user_id,
-            cluster_id,
-            postgresql_where=(role == ClusterRole.admin),
-        ),
-    )
-
-
-class ClusterPageAssociation(Base):
-    """
-    pages related to clusters
-    """
-
-    __tablename__ = "cluster_page_associations"
-    __table_args__ = (UniqueConstraint("page_id", "cluster_id"),)
-
-    id = Column(BigInteger, primary_key=True)
-
-    page_id = Column(ForeignKey("pages.id"), nullable=False, index=True)
-    cluster_id = Column(ForeignKey("clusters.id"), nullable=False, index=True)
-
-    page = relationship("Page", backref="cluster_page_associations")
-    cluster = relationship("Cluster", backref="cluster_page_associations")
-
-
-class PageType(enum.Enum):
-    main_page = enum.auto()
-    place = enum.auto()
-    guide = enum.auto()
-
-
-class Page(Base):
-    """
-    similar to a wiki page about a community, POI or guide
-    """
-
-    __tablename__ = "pages"
-
-    id = Column(BigInteger, communities_seq, primary_key=True, server_default=communities_seq.next_value())
-
-    parent_node_id = Column(ForeignKey("nodes.id"), nullable=False, index=True)
-    type = Column(Enum(PageType), nullable=False)
-    creator_user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    owner_user_id = Column(ForeignKey("users.id"), nullable=True, index=True)
-    owner_cluster_id = Column(ForeignKey("clusters.id"), nullable=True, index=True)
-
-    thread_id = Column(ForeignKey("threads.id"), nullable=False, unique=True)
-
-    parent_node = relationship("Node", backref="child_pages", remote_side="Node.id", foreign_keys="Page.parent_node_id")
-
-    thread = relationship("Thread", backref="page", uselist=False)
-    creator_user = relationship("User", backref="created_pages", foreign_keys="Page.creator_user_id")
-    owner_user = relationship("User", backref="owned_pages", foreign_keys="Page.owner_user_id")
-    owner_cluster = relationship(
-        "Cluster", backref=backref("owned_pages", lazy="dynamic"), uselist=False, foreign_keys="Page.owner_cluster_id"
-    )
-
-    editors = relationship("User", secondary="page_versions", viewonly=True)
-
-    __table_args__ = (
-        # Only one of owner_user and owner_cluster should be set
-        CheckConstraint(
-            "(owner_user_id IS NULL) <> (owner_cluster_id IS NULL)",
-            name="one_owner",
-        ),
-        # Only clusters can own main pages
-        CheckConstraint(
-            "NOT (owner_cluster_id IS NULL AND type = 'main_page')",
-            name="main_page_owned_by_cluster",
-        ),
-        # Each cluster can have at most one main page
-        Index(
-            "ix_pages_owner_cluster_id_type",
-            owner_cluster_id,
-            type,
-            unique=True,
-            postgresql_where=(type == PageType.main_page),
-        ),
-    )
-
-    def __repr__(self):
-        return f"Page({self.id=})"
-
-
-class PageVersion(Base):
-    """
-    version of page content
-    """
-
-    __tablename__ = "page_versions"
-
-    id = Column(BigInteger, primary_key=True)
-
-    page_id = Column(ForeignKey("pages.id"), nullable=False, index=True)
-    editor_user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    title = Column(String, nullable=False)
-    content = Column(String, nullable=False)  # CommonMark without images
-    photo_key = Column(ForeignKey("uploads.key"), nullable=True)
-    geom = Column(Geometry(geometry_type="POINT", srid=4326), nullable=True)
-    # the human-readable address
-    address = Column(String, nullable=True)
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    slug = column_property(func.slugify(title))
-
-    page = relationship("Page", backref="versions", order_by="PageVersion.id")
-    editor_user = relationship("User", backref="edited_pages")
-    photo = relationship("Upload")
-
-    __table_args__ = (
-        # Geom and address must either both be null or both be set
-        CheckConstraint(
-            "(geom IS NULL) = (address IS NULL)",
-            name="geom_iff_address",
-        ),
-    )
-
-    @property
-    def coordinates(self):
-        # returns (lat, lng) or None
-        return get_coordinates(self.geom)
-
-    def __repr__(self):
-        return f"PageVersion({self.id=}, {self.page_id=})"
-
-
-class ClusterEventAssociation(Base):
-    """
-    events related to clusters
-    """
-
-    __tablename__ = "cluster_event_associations"
-    __table_args__ = (UniqueConstraint("event_id", "cluster_id"),)
-
-    id = Column(BigInteger, primary_key=True)
-
-    event_id = Column(ForeignKey("events.id"), nullable=False, index=True)
-    cluster_id = Column(ForeignKey("clusters.id"), nullable=False, index=True)
-
-    event = relationship("Event", backref="cluster_event_associations")
-    cluster = relationship("Cluster", backref="cluster_event_associations")
-
-
-class Event(Base):
-    """
-    An event is compose of two parts:
-
-    * An event template (Event)
-    * An occurrence (EventOccurrence)
-
-    One-off events will have one of each; repeating events will have one Event,
-    multiple EventOccurrences, one for each time the event happens.
-    """
-
-    __tablename__ = "events"
-
-    id = Column(BigInteger, communities_seq, primary_key=True, server_default=communities_seq.next_value())
-    parent_node_id = Column(ForeignKey("nodes.id"), nullable=False, index=True)
-
-    title = Column(String, nullable=False)
-
-    slug = column_property(func.slugify(title))
-
-    creator_user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    owner_user_id = Column(ForeignKey("users.id"), nullable=True, index=True)
-    owner_cluster_id = Column(ForeignKey("clusters.id"), nullable=True, index=True)
-    thread_id = Column(ForeignKey("threads.id"), nullable=False, unique=True)
-
-    parent_node = relationship(
-        "Node", backref="child_events", remote_side="Node.id", foreign_keys="Event.parent_node_id"
-    )
-    thread = relationship("Thread", backref="event", uselist=False)
-    subscribers = relationship(
-        "User", backref="subscribed_events", secondary="event_subscriptions", lazy="dynamic", viewonly=True
-    )
-    organizers = relationship(
-        "User", backref="organized_events", secondary="event_organizers", lazy="dynamic", viewonly=True
-    )
-    creator_user = relationship("User", backref="created_events", foreign_keys="Event.creator_user_id")
-    owner_user = relationship("User", backref="owned_events", foreign_keys="Event.owner_user_id")
-    owner_cluster = relationship(
-        "Cluster",
-        backref=backref("owned_events", lazy="dynamic"),
-        uselist=False,
-        foreign_keys="Event.owner_cluster_id",
-    )
-
-    __table_args__ = (
-        # Only one of owner_user and owner_cluster should be set
-        CheckConstraint(
-            "(owner_user_id IS NULL) <> (owner_cluster_id IS NULL)",
-            name="one_owner",
-        ),
-    )
-
-
-class EventOccurrence(Base):
-    __tablename__ = "event_occurrences"
-
-    id = Column(BigInteger, communities_seq, primary_key=True, server_default=communities_seq.next_value())
-    event_id = Column(ForeignKey("events.id"), nullable=False, index=True)
-
-    # the user that created this particular occurrence of a repeating event (same as event.creator_user_id if single event)
-    creator_user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    content = Column(String, nullable=False)  # CommonMark without images
-    photo_key = Column(ForeignKey("uploads.key"), nullable=True)
-
-    is_cancelled = Column(Boolean, nullable=False, default=False, server_default=expression.false())
-    is_deleted = Column(Boolean, nullable=False, default=False, server_default=expression.false())
-
-    # a null geom is an online-only event
-    geom = Column(Geometry(geometry_type="POINT", srid=4326), nullable=True)
-    # physical address, iff geom is not null
-    address = Column(String, nullable=True)
-    # videoconferencing link, etc, must be specified if no geom, otherwise optional
-    link = Column(String, nullable=True)
-
-    timezone = "Etc/UTC"
-
-    # time during which the event takes place; this is a range type (instead of separate start+end times) which
-    # simplifies database constraints, etc
-    during = Column(TSTZRANGE, nullable=False)
-
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    last_edited = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    creator_user = relationship(
-        "User", backref="created_event_occurrences", foreign_keys="EventOccurrence.creator_user_id"
-    )
-    event = relationship(
-        "Event",
-        backref=backref("occurrences", lazy="dynamic"),
-        remote_side="Event.id",
-        foreign_keys="EventOccurrence.event_id",
-    )
-
-    photo = relationship("Upload")
-
-    __table_args__ = (
-        # Geom and address go together
-        CheckConstraint(
-            # geom and address are either both null or neither of them are null
-            "(geom IS NULL) = (address IS NULL)",
-            name="geom_iff_address",
-        ),
-        # Online-only events need a link, note that online events may also have a link
-        CheckConstraint(
-            # exactly oen of geom or link is non-null
-            "(geom IS NULL) <> (link IS NULL)",
-            name="link_or_geom",
-        ),
-        # Can't have overlapping occurrences in the same Event
-        ExcludeConstraint(("event_id", "="), ("during", "&&"), name="event_occurrences_event_id_during_excl"),
-    )
-
-    @property
-    def coordinates(self):
-        # returns (lat, lng) or None
-        return get_coordinates(self.geom)
-
-    @hybrid_property
-    def start_time(self):
-        return self.during.lower
-
-    @start_time.expression
-    def start_time(cls):
-        return func.lower(cls.during)
-
-    @hybrid_property
-    def end_time(self):
-        return self.during.upper
-
-    @end_time.expression
-    def end_time(cls):
-        return func.upper(cls.during)
-
-
-class EventSubscription(Base):
-    """
-    Users' subscriptions to events
-    """
-
-    __tablename__ = "event_subscriptions"
-    __table_args__ = (UniqueConstraint("event_id", "user_id"),)
-
-    id = Column(BigInteger, primary_key=True)
-
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    event_id = Column(ForeignKey("events.id"), nullable=False, index=True)
-    joined = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    user = relationship("User")
-    event = relationship("Event")
-
-
-class EventOrganizer(Base):
-    """
-    Organizers for events
-    """
-
-    __tablename__ = "event_organizers"
-    __table_args__ = (UniqueConstraint("event_id", "user_id"),)
-
-    id = Column(BigInteger, primary_key=True)
-
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    event_id = Column(ForeignKey("events.id"), nullable=False, index=True)
-    joined = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    user = relationship("User")
-    event = relationship("Event")
-
-
-class AttendeeStatus(enum.Enum):
-    going = enum.auto()
-    maybe = enum.auto()
-
-
-class EventOccurrenceAttendee(Base):
-    """
-    Attendees for events
-    """
-
-    __tablename__ = "event_occurrence_attendees"
-    __table_args__ = (UniqueConstraint("occurrence_id", "user_id"),)
-
-    id = Column(BigInteger, primary_key=True)
-
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    occurrence_id = Column(ForeignKey("event_occurrences.id"), nullable=False, index=True)
-    responded = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    attendee_status = Column(Enum(AttendeeStatus), nullable=False)
-
-    user = relationship("User")
-    occurrence = relationship("EventOccurrence", backref=backref("attendances", lazy="dynamic"))
-
-    reminder_sent = Column(Boolean, nullable=False, default=False, server_default=expression.false())
-
-
-class EventCommunityInviteRequest(Base):
-    """
-    Requests to send out invitation notifications/emails to the community for a given event occurrence
-    """
-
-    __tablename__ = "event_community_invite_requests"
-
-    id = Column(BigInteger, primary_key=True)
-
-    occurrence_id = Column(ForeignKey("event_occurrences.id"), nullable=False, index=True)
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    decided = Column(DateTime(timezone=True), nullable=True)
-    decided_by_user_id = Column(ForeignKey("users.id"), nullable=True)
-    approved = Column(Boolean, nullable=True)
-
-    occurrence = relationship("EventOccurrence", backref=backref("community_invite_requests", lazy="dynamic"))
-    user = relationship("User", foreign_keys="EventCommunityInviteRequest.user_id")
-
-    __table_args__ = (
-        # each user can only request once
-        UniqueConstraint("occurrence_id", "user_id"),
-        # each event can only have one notification sent out
-        Index(
-            "ix_event_community_invite_requests_unique",
-            occurrence_id,
-            unique=True,
-            postgresql_where=and_(approved.is_not(None), approved == True),
-        ),
-        # decided and approved ought to be null simultaneously
-        CheckConstraint(
-            "((decided IS NULL) AND (decided_by_user_id IS NULL) AND (approved IS NULL)) OR \
-             ((decided IS NOT NULL) AND (decided_by_user_id IS NOT NULL) AND (approved IS NOT NULL))",
-            name="decided_approved",
-        ),
-    )
-
-
-class ClusterDiscussionAssociation(Base):
-    """
-    discussions related to clusters
-    """
-
-    __tablename__ = "cluster_discussion_associations"
-    __table_args__ = (UniqueConstraint("discussion_id", "cluster_id"),)
-
-    id = Column(BigInteger, primary_key=True)
-
-    discussion_id = Column(ForeignKey("discussions.id"), nullable=False, index=True)
-    cluster_id = Column(ForeignKey("clusters.id"), nullable=False, index=True)
-
-    discussion = relationship("Discussion", backref="cluster_discussion_associations")
-    cluster = relationship("Cluster", backref="cluster_discussion_associations")
-
-
-class Discussion(Base):
-    """
-    forum board
-    """
-
-    __tablename__ = "discussions"
-
-    id = Column(BigInteger, communities_seq, primary_key=True, server_default=communities_seq.next_value())
-
-    title = Column(String, nullable=False)
-    content = Column(String, nullable=False)
-    thread_id = Column(ForeignKey("threads.id"), nullable=False, unique=True)
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    creator_user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    owner_cluster_id = Column(ForeignKey("clusters.id"), nullable=False, index=True)
-
-    slug = column_property(func.slugify(title))
-
-    thread = relationship("Thread", backref="discussion", uselist=False)
-
-    subscribers = relationship("User", backref="discussions", secondary="discussion_subscriptions", viewonly=True)
-
-    creator_user = relationship("User", backref="created_discussions", foreign_keys="Discussion.creator_user_id")
-    owner_cluster = relationship("Cluster", backref=backref("owned_discussions", lazy="dynamic"), uselist=False)
-
-
-class DiscussionSubscription(Base):
-    """
-    users subscriptions to discussions
-    """
-
-    __tablename__ = "discussion_subscriptions"
-    __table_args__ = (UniqueConstraint("discussion_id", "user_id"),)
-
-    id = Column(BigInteger, primary_key=True)
-
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-    discussion_id = Column(ForeignKey("discussions.id"), nullable=False, index=True)
-    joined = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    left = Column(DateTime(timezone=True), nullable=True)
-
-    user = relationship("User", backref="discussion_subscriptions")
-    discussion = relationship("Discussion", backref="discussion_subscriptions")
-
-
-class Thread(Base):
-    """
-    Thread
-    """
-
-    __tablename__ = "threads"
-
-    id = Column(BigInteger, primary_key=True)
-
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    deleted = Column(DateTime(timezone=True), nullable=True)
-
-
-class Comment(Base):
-    """
-    Comment
-    """
-
-    __tablename__ = "comments"
-
-    id = Column(BigInteger, primary_key=True)
-
-    thread_id = Column(ForeignKey("threads.id"), nullable=False, index=True)
-    author_user_id = Column(ForeignKey("users.id"), nullable=False)
-    content = Column(String, nullable=False)  # CommonMark without images
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    deleted = Column(DateTime(timezone=True), nullable=True)
-
-    thread = relationship("Thread", backref="comments")
-
-
-class Reply(Base):
-    """
-    Reply
-    """
-
-    __tablename__ = "replies"
-
-    id = Column(BigInteger, primary_key=True)
-
-    comment_id = Column(ForeignKey("comments.id"), nullable=False, index=True)
-    author_user_id = Column(ForeignKey("users.id"), nullable=False)
-    content = Column(String, nullable=False)  # CommonMark without images
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    deleted = Column(DateTime(timezone=True), nullable=True)
-
-    comment = relationship("Comment", backref="replies")
-
-
-class BackgroundJobState(enum.Enum):
-    # job is fresh, waiting to be picked off the queue
-    pending = enum.auto()
-    # job complete
-    completed = enum.auto()
-    # error occured, will be retried
-    error = enum.auto()
-    # failed too many times, not retrying anymore
-    failed = enum.auto()
-
-
-class BackgroundJob(Base):
-    """
-    This table implements a queue of background jobs.
-    """
-
-    __tablename__ = "background_jobs"
-
-    id = Column(BigInteger, primary_key=True)
-
-    # used to discern which function should be triggered to service it
-    job_type = Column(String, nullable=False)
-    state = Column(Enum(BackgroundJobState), nullable=False, default=BackgroundJobState.pending)
-
-    # time queued
-    queued = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    # time at which we may next attempt it, for implementing exponential backoff
-    next_attempt_after = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    # used to count number of retries for failed jobs
-    try_count = Column(Integer, nullable=False, default=0)
-
-    max_tries = Column(Integer, nullable=False, default=5)
-
-    # higher is more important
-    priority = Column(Integer, nullable=False, server_default=text("10"))
-
-    # protobuf encoded job payload
-    payload = Column(Binary, nullable=False)
-
-    # if the job failed, we write that info here
-    failure_info = Column(String, nullable=True)
-
-    __table_args__ = (
-        # used in looking up background jobs to attempt
-        # create index on background_jobs(priority desc, next_attempt_after, (max_tries - try_count)) where state = 'pending' OR state = 'error';
-        Index(
-            "ix_background_jobs_lookup",
-            priority.desc(),
-            next_attempt_after,
-            (max_tries - try_count),
-            postgresql_where=((state == BackgroundJobState.pending) | (state == BackgroundJobState.error)),
-        ),
-    )
-
-    @hybrid_property
-    def ready_for_retry(self):
-        return (
-            (self.next_attempt_after <= func.now())
-            & (self.try_count < self.max_tries)
-            & ((self.state == BackgroundJobState.pending) | (self.state == BackgroundJobState.error))
-        )
-
-    def __repr__(self):
-        return f"BackgroundJob(id={self.id}, job_type={self.job_type}, state={self.state}, next_attempt_after={self.next_attempt_after}, try_count={self.try_count}, failure_info={self.failure_info})"
-
-
-class NotificationDeliveryType(enum.Enum):
-    # send push notification to mobile/web
-    push = enum.auto()
-    # send individual email immediately
-    email = enum.auto()
-    # send in digest
-    digest = enum.auto()
-
-
-dt = NotificationDeliveryType
-nd = notification_data_pb2
-
-dt_sec = [dt.email, dt.push]
-dt_all = [dt.email, dt.push, dt.digest]
-
-
-class NotificationTopicAction(enum.Enum):
-    def __init__(self, topic_action, defaults, user_editable, data_type):
-        self.topic, self.action = topic_action.split(":")
-        self.defaults = defaults
-        # for now user editable == not a security notification
-        self.user_editable = user_editable
-
-        self.data_type = data_type
-
-    def unpack(self):
-        return self.topic, self.action
-
-    @property
-    def display(self):
-        return f"{self.topic}:{self.action}"
-
-    def __str__(self):
-        return self.display
-
-    # topic, action, default delivery types
-    friend_request__create = ("friend_request:create", dt_all, True, nd.FriendRequestCreate)
-    friend_request__accept = ("friend_request:accept", dt_all, True, nd.FriendRequestAccept)
-
-    # host requests
-    host_request__create = ("host_request:create", dt_all, True, nd.HostRequestCreate)
-    host_request__accept = ("host_request:accept", dt_all, True, nd.HostRequestAccept)
-    host_request__reject = ("host_request:reject", dt_all, True, nd.HostRequestReject)
-    host_request__confirm = ("host_request:confirm", dt_all, True, nd.HostRequestConfirm)
-    host_request__cancel = ("host_request:cancel", dt_all, True, nd.HostRequestCancel)
-    host_request__message = ("host_request:message", [dt.push, dt.digest], True, nd.HostRequestMessage)
-    host_request__missed_messages = ("host_request:missed_messages", [dt.email], True, nd.HostRequestMissedMessages)
-    host_request__reminder = ("host_request:reminder", dt_all, True, nd.HostRequestReminder)
-
-    activeness__probe = ("activeness:probe", dt_sec, False, nd.ActivenessProbe)
-
-    # you receive a friend ref
-    reference__receive_friend = ("reference:receive_friend", dt_all, True, nd.ReferenceReceiveFriend)
-    # you receive a reference from ... the host
-    reference__receive_hosted = ("reference:receive_hosted", dt_all, True, nd.ReferenceReceiveHostRequest)
-    # ... the surfer
-    reference__receive_surfed = ("reference:receive_surfed", dt_all, True, nd.ReferenceReceiveHostRequest)
-
-    # you hosted
-    reference__reminder_hosted = ("reference:reminder_hosted", dt_all, True, nd.ReferenceReminder)
-    # you surfed
-    reference__reminder_surfed = ("reference:reminder_surfed", dt_all, True, nd.ReferenceReminder)
-
-    badge__add = ("badge:add", [dt.push, dt.digest], True, nd.BadgeAdd)
-    badge__remove = ("badge:remove", [dt.push, dt.digest], True, nd.BadgeRemove)
-
-    # group chats
-    chat__message = ("chat:message", [dt.push, dt.digest], True, nd.ChatMessage)
-    chat__missed_messages = ("chat:missed_messages", [dt.email], True, nd.ChatMissedMessages)
-
-    # events
-    # approved by mods
-    event__create_approved = ("event:create_approved", dt_all, True, nd.EventCreate)
-    # any user creates any event, default to no notifications
-    event__create_any = ("event:create_any", [], True, nd.EventCreate)
-    event__update = ("event:update", dt_all, True, nd.EventUpdate)
-    event__cancel = ("event:cancel", dt_all, True, nd.EventCancel)
-    event__delete = ("event:delete", dt_all, True, nd.EventDelete)
-    event__invite_organizer = ("event:invite_organizer", dt_all, True, nd.EventInviteOrganizer)
-    event__reminder = ("event:reminder", dt_all, True, nd.EventReminder)
-    # toplevel comment on an event
-    event__comment = ("event:comment", dt_all, True, nd.EventComment)
-
-    # discussion created
-    discussion__create = ("discussion:create", [dt.digest], True, nd.DiscussionCreate)
-    # someone comments on your discussion
-    discussion__comment = ("discussion:comment", dt_all, True, nd.DiscussionComment)
-
-    # someone responds to any of your top-level comment across the platform
-    thread__reply = ("thread:reply", dt_all, True, nd.ThreadReply)
-
-    # account settings
-    password__change = ("password:change", dt_sec, False, empty_pb2.Empty)
-    email_address__change = ("email_address:change", dt_sec, False, nd.EmailAddressChange)
-    email_address__verify = ("email_address:verify", dt_sec, False, empty_pb2.Empty)
-    phone_number__change = ("phone_number:change", dt_sec, False, nd.PhoneNumberChange)
-    phone_number__verify = ("phone_number:verify", dt_sec, False, nd.PhoneNumberVerify)
-    # reset password
-    password_reset__start = ("password_reset:start", dt_sec, False, nd.PasswordResetStart)
-    password_reset__complete = ("password_reset:complete", dt_sec, False, empty_pb2.Empty)
-
-    # account deletion
-    account_deletion__start = ("account_deletion:start", dt_sec, False, nd.AccountDeletionStart)
-    # no more pushing to do
-    account_deletion__complete = ("account_deletion:complete", dt_sec, False, nd.AccountDeletionComplete)
-    # undeleted
-    account_deletion__recovered = ("account_deletion:recovered", dt_sec, False, empty_pb2.Empty)
-
-    # admin actions
-    gender__change = ("gender:change", dt_sec, False, nd.GenderChange)
-    birthdate__change = ("birthdate:change", dt_sec, False, nd.BirthdateChange)
-    api_key__create = ("api_key:create", dt_sec, False, nd.ApiKeyCreate)
-
-    donation__received = ("donation:received", dt_sec, True, nd.DonationReceived)
-
-    onboarding__reminder = ("onboarding:reminder", dt_sec, True, empty_pb2.Empty)
-
-    modnote__create = ("modnote:create", dt_sec, False, empty_pb2.Empty)
-
-    verification__sv_fail = ("verification:sv_fail", dt_sec, False, nd.VerificationSVFail)
-    verification__sv_success = ("verification:sv_success", dt_sec, False, empty_pb2.Empty)
-
-    # general announcements
-    general__new_blog_post = ("general:new_blog_post", [dt.push, dt.digest], True, nd.GeneralNewBlogPost)
-
-
-class NotificationPreference(Base):
-    __tablename__ = "notification_preferences"
-
-    id = Column(BigInteger, primary_key=True)
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-
-    topic_action = Column(Enum(NotificationTopicAction), nullable=False)
-    delivery_type = Column(Enum(NotificationDeliveryType), nullable=False)
-    deliver = Column(Boolean, nullable=False)
-
-    user = relationship("User", foreign_keys="NotificationPreference.user_id")
-
-    __table_args__ = (UniqueConstraint("user_id", "topic_action", "delivery_type"),)
-
-
-class Notification(Base):
-    """
-    Table for accumulating notifications until it is time to send email digest
-    """
-
-    __tablename__ = "notifications"
-
-    id = Column(BigInteger, primary_key=True)
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    # recipient user id
-    user_id = Column(ForeignKey("users.id"), nullable=False)
-
-    topic_action = Column(Enum(NotificationTopicAction), nullable=False)
-    key = Column(String, nullable=False)
-
-    data = Column(Binary, nullable=False)
-
-    # whether the user has marked this notification as seen or not
-    is_seen = Column(Boolean, nullable=False, server_default=expression.false())
-
-    user = relationship("User", foreign_keys="Notification.user_id")
-
-    __table_args__ = (
-        # used in looking up which notifications need delivery
-        Index(
-            "ix_notifications_created",
-            created,
-        ),
-        # Fast lookup for unseen notification count
-        Index(
-            "ix_notifications_unseen",
-            user_id,
-            topic_action,
-            postgresql_where=(is_seen == False),
-        ),
-        # Fast lookup for latest notifications
-        Index(
-            "ix_notifications_latest",
-            user_id,
-            id.desc(),
-            topic_action,
-        ),
-    )
-
-    @property
-    def topic(self):
-        return self.topic_action.topic
-
-    @property
-    def action(self):
-        return self.topic_action.action
-
-
-class NotificationDelivery(Base):
-    __tablename__ = "notification_deliveries"
-
-    id = Column(BigInteger, primary_key=True)
-    notification_id = Column(ForeignKey("notifications.id"), nullable=False, index=True)
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    delivered = Column(DateTime(timezone=True), nullable=True)
-    read = Column(DateTime(timezone=True), nullable=True)
-    # todo: enum of "phone, web, digest"
-    delivery_type = Column(Enum(NotificationDeliveryType), nullable=False)
-    # todo: device id
-    # todo: receipt id, etc
-    notification = relationship("Notification", foreign_keys="NotificationDelivery.notification_id")
-
-    __table_args__ = (
-        UniqueConstraint("notification_id", "delivery_type"),
-        # used in looking up which notifications need delivery
-        Index(
-            "ix_notification_deliveries_delivery_type",
-            delivery_type,
-            postgresql_where=(delivered != None),
-        ),
-        Index(
-            "ix_notification_deliveries_dt_ni_dnull",
-            delivery_type,
-            notification_id,
-            delivered == None,
-        ),
-    )
-
-
-class PushNotificationSubscription(Base):
-    __tablename__ = "push_notification_subscriptions"
-
-    id = Column(BigInteger, primary_key=True)
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    # which user this is connected to
-    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
-
-    # these come from https://developer.mozilla.org/en-US/docs/Web/API/PushSubscription
-    # the endpoint
-    endpoint = Column(String, nullable=False)
-    # the "auth" key
-    auth_key = Column(Binary, nullable=False)
-    # the "p256dh" key
-    p256dh_key = Column(Binary, nullable=False)
-
-    full_subscription_info = Column(String, nullable=False)
-
-    # the browse user-agent, so we can tell the user what browser notifications are going to
-    user_agent = Column(String, nullable=True)
-
-    # when it was disabled
-    disabled_at = Column(DateTime(timezone=True), nullable=False, server_default=DATETIME_INFINITY.isoformat())
-
-    user = relationship("User")
-
-
-class PushNotificationDeliveryAttempt(Base):
-    __tablename__ = "push_notification_delivery_attempt"
-
-    id = Column(BigInteger, primary_key=True)
-    time = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    push_notification_subscription_id = Column(
-        ForeignKey("push_notification_subscriptions.id"), nullable=False, index=True
-    )
-
-    success = Column(Boolean, nullable=False)
-    # the HTTP status code, 201 is success
-    status_code = Column(Integer, nullable=False)
-
-    # can be null if it was a success
-    response = Column(String, nullable=True)
-
-    push_notification_subscription = relationship("PushNotificationSubscription")
-
-
-class Language(Base):
-    """
-    Table of allowed languages (a subset of ISO639-3)
-    """
-
-    __tablename__ = "languages"
-
-    # ISO639-3 language code, in lowercase, e.g. fin, eng
-    code = Column(String(3), primary_key=True)
-
-    # the english name
-    name = Column(String, nullable=False, unique=True)
-
-
-class Region(Base):
-    """
-    Table of regions
-    """
-
-    __tablename__ = "regions"
-
-    # iso 3166-1 alpha3 code in uppercase, e.g. FIN, USA
-    code = Column(String(3), primary_key=True)
-
-    # the name, e.g. Finland, United States
-    # this is the display name in English, should be the "common name", not "Republic of Finland"
-    name = Column(String, nullable=False, unique=True)
-
-
-class UserBlock(Base):
-    """
-    Table of blocked users
-    """
-
-    __tablename__ = "user_blocks"
-
-    id = Column(BigInteger, primary_key=True)
-
-    blocking_user_id = Column(ForeignKey("users.id"), nullable=False)
-    blocked_user_id = Column(ForeignKey("users.id"), nullable=False)
-    time_blocked = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    blocking_user = relationship("User", foreign_keys="UserBlock.blocking_user_id")
-    blocked_user = relationship("User", foreign_keys="UserBlock.blocked_user_id")
-
-    __table_args__ = (
-        UniqueConstraint("blocking_user_id", "blocked_user_id"),
-        Index("ix_user_blocks_blocking_user_id", blocking_user_id, blocked_user_id),
-        Index("ix_user_blocks_blocked_user_id", blocked_user_id, blocking_user_id),
-    )
-
-
-class APICall(Base):
-    """
-    API call logs
-    """
-
-    __tablename__ = "api_calls"
-    __table_args__ = {"schema": "logging"}
-
-    id = Column(BigInteger, primary_key=True)
-
-    # whether the call was made using an api key or session cookies
-    is_api_key = Column(Boolean, nullable=False, server_default=expression.false())
-
-    # backend version (normally e.g. develop-31469e3), allows us to figure out which proto definitions were used
-    # note that `default` is a python side default, not hardcoded into DB schema
-    version = Column(String, nullable=False, default=config["VERSION"])
-
-    # approximate time of the call
-    time = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    # the method call name, e.g. "/org.couchers.api.core.API/ListFriends"
-    method = Column(String, nullable=False)
-
-    # gRPC status code name, e.g. FAILED_PRECONDITION, None if success
-    status_code = Column(String, nullable=True)
-
-    # handler duration (excluding serialization, etc)
-    duration = Column(Float, nullable=False)
-
-    # user_id of caller, None means not logged in
-    user_id = Column(BigInteger, nullable=True)
-
-    # sanitized request bytes
-    request = Column(Binary, nullable=True)
-
-    # sanitized response bytes
-    response = Column(Binary, nullable=True)
-
-    # whether response bytes have been truncated
-    response_truncated = Column(Boolean, nullable=False, server_default=expression.false())
-
-    # the exception traceback, if any
-    traceback = Column(String, nullable=True)
-
-    # human readable perf report
-    perf_report = Column(String, nullable=True)
-
-    # details of the browser, if available
-    ip_address = Column(String, nullable=True)
-    user_agent = Column(String, nullable=True)
-
-
-class AccountDeletionReason(Base):
-    __tablename__ = "account_deletion_reason"
-
-    id = Column(BigInteger, primary_key=True)
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    user_id = Column(ForeignKey("users.id"), nullable=False)
-    reason = Column(String, nullable=True)
-
-    user = relationship("User")
-
-
-class ModerationUserList(Base):
-    """
-    Represents a list of users listed together by a moderator
-    """
-
-    __tablename__ = "moderation_user_lists"
-
-    id = Column(BigInteger, primary_key=True)
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-
-    # Relationships
-    users = relationship("User", secondary="moderation_user_list_members", back_populates="moderation_user_lists")
-
-
-class ModerationUserListMember(Base):
-    """
-    Association table for many-to-many relationship between users and moderation_user_lists
-    """
-
-    __tablename__ = "moderation_user_list_members"
-
-    user_id = Column(ForeignKey("users.id"), primary_key=True)
-    moderation_list_id = Column(ForeignKey("moderation_user_lists.id"), primary_key=True)
-
-    __table_args__ = (UniqueConstraint("user_id", "moderation_list_id"),)
-
-
-class AntiBotLog(Base):
-    __tablename__ = "antibot_logs"
-
-    id = Column(BigInteger, primary_key=True)
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    user_id = Column(ForeignKey("users.id"), nullable=True)
-
-    ip_address = Column(String, nullable=True)
-    user_agent = Column(String, nullable=True)
-
-    action = Column(String, nullable=False)
-    token = Column(String, nullable=False)
-
-    score = Column(Float, nullable=False)
-    provider_data = Column(JSON, nullable=False)
-
-
-class RateLimitAction(enum.Enum):
-    """Possible user actions which can be rate limited."""
-
-    host_request = "host request"
-    friend_request = "friend request"
-    chat_initiation = "chat initiation"
-
-
-class RateLimitViolation(Base):
-    __tablename__ = "rate_limit_violations"
-
-    id = Column(BigInteger, primary_key=True)
-    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
-    user_id = Column(ForeignKey("users.id"), nullable=False)
-    action = Column(Enum(RateLimitAction), nullable=False)
-    is_hard_limit = Column(Boolean, nullable=False)
-
-    user = relationship("User")
-
-    __table_args__ = (
-        # Fast lookup for rate limits in interval
-        Index("ix_rate_limits_by_user", user_id, action, is_hard_limit, created),
-    )
-
-
-class Volunteer(Base):
-    __tablename__ = "volunteers"
-
-    id = Column(BigInteger, primary_key=True)
-    user_id = Column(ForeignKey("users.id"), nullable=False, unique=True)
-
-    display_name = Column(String, nullable=True)
-    display_location = Column(String, nullable=True)
-
-    role = Column(String, nullable=False)
-
-    # custom sort order on team page, sorted ascending
-    sort_key = Column(Float, nullable=True)
-
-    started_volunteering = Column(Date, nullable=False, server_default=text("CURRENT_DATE"))
-    stopped_volunteering = Column(Date, nullable=True, default=None)
-
-    link_type = Column(String, nullable=True)
-    link_text = Column(String, nullable=True)
-    link_url = Column(String, nullable=True)
-
-    show_on_team_page = Column(Boolean, nullable=False, server_default=expression.true())
-
-    __table_args__ = (
-        # Link type, text, url should all be null or all not be null
-        CheckConstraint(
-            "(link_type IS NULL) = (link_text IS NULL) AND (link_type IS NULL) = (link_url IS NULL)",
-            name="link_type_text",
-        ),
-    )
