+++ app/backend/src/couchers/models/__init__.py
@@ -1,0 +1,18 @@
+from .activeness_probe import *  # noqa: F401,F403
+from .auth import *  # noqa: F401,F403
+from .background_jobs import *  # noqa: F401,F403
+from .base import *  # noqa: F401,F403
+from .clusters import *  # noqa: F401,F403
+from .conversations import *  # noqa: F401,F403
+from .discussions import *  # noqa: F401,F403
+from .donations import *  # noqa: F401,F403
+from .events import *  # noqa: F401,F403
+from .host_requests import *  # noqa: F401,F403
+from .logging import *  # noqa: F401,F403
+from .mod_note import *  # noqa: F401,F403
+from .notifications import *  # noqa: F401,F403
+from .rest import *  # noqa: F401,F403
+from .static import *  # noqa: F401,F403
+from .uploads import *  # noqa: F401,F403
+from .users import *  # noqa: F401,F403
+from .verification import *  # noqa: F401,F403
