+++ app/backend/src/couchers/models/base.py
@@ -1,0 +1,19 @@
+from sqlalchemy import MetaData, Sequence
+from sqlalchemy.orm import DeclarativeBase
+
+meta = MetaData(
+    naming_convention={
+        "ix": "ix_%(column_0_label)s",
+        "uq": "uq_%(table_name)s_%(column_0_name)s",
+        "ck": "ck_%(table_name)s_%(constraint_name)s",
+        "fk": "fk_%(table_name)s_%(column_0_name)s_%(referred_table_name)s",
+        "pk": "pk_%(table_name)s",
+    }
+)
+
+
+class Base(DeclarativeBase):
+    metadata = meta
+
+
+communities_seq = Sequence("communities_seq")
