+++ app/backend/src/couchers/models/conversations.py
@@ -1,0 +1,162 @@
+import enum
+
+from sqlalchemy import BigInteger, Boolean, Column, DateTime, Enum, ForeignKey, String, func
+from sqlalchemy.ext.hybrid import hybrid_property
+from sqlalchemy.orm import backref, relationship
+
+from couchers.constants import DATETIME_INFINITY, DATETIME_MINUS_INFINITY
+from couchers.models.base import Base
+from couchers.models.host_requests import HostRequestStatus
+from couchers.utils import now
+
+
+class Conversation(Base):
+    """
+    Conversation brings together the different types of message/conversation types
+    """
+
+    __tablename__ = "conversations"
+
+    id = Column(BigInteger, primary_key=True)
+    # timezone should always be UTC
+    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
+
+    def __repr__(self):
+        return f"Conversation(id={self.id}, created={self.created})"
+
+
+class GroupChat(Base):
+    """
+    Group chat
+    """
+
+    __tablename__ = "group_chats"
+
+    conversation_id = Column("id", ForeignKey("conversations.id"), nullable=False, primary_key=True)
+
+    title = Column(String, nullable=True)
+    only_admins_invite = Column(Boolean, nullable=False, default=True)
+    creator_id = Column(ForeignKey("users.id"), nullable=False, index=True)
+    is_dm = Column(Boolean, nullable=False)
+
+    conversation = relationship("Conversation", backref="group_chat")
+    creator = relationship("User", backref="created_group_chats")
+
+    def __repr__(self):
+        return f"GroupChat(conversation={self.conversation}, title={self.title or 'None'}, only_admins_invite={self.only_admins_invite}, creator={self.creator}, is_dm={self.is_dm})"
+
+
+class GroupChatRole(enum.Enum):
+    admin = enum.auto()
+    participant = enum.auto()
+
+
+class GroupChatSubscription(Base):
+    """
+    The recipient of a thread and information about when they joined/left/etc.
+    """
+
+    __tablename__ = "group_chat_subscriptions"
+    id = Column(BigInteger, primary_key=True)
+
+    # TODO: DB constraint on only one user+group_chat combo at a given time
+    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
+    group_chat_id = Column(ForeignKey("group_chats.id"), nullable=False, index=True)
+
+    # timezones should always be UTC
+    joined = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
+    left = Column(DateTime(timezone=True), nullable=True)
+
+    role = Column(Enum(GroupChatRole), nullable=False)
+
+    last_seen_message_id = Column(BigInteger, nullable=False, default=0)
+
+    # when this chat is muted until, DATETIME_INFINITY for "forever"
+    muted_until = Column(DateTime(timezone=True), nullable=False, server_default=DATETIME_MINUS_INFINITY.isoformat())
+
+    user = relationship("User", backref="group_chat_subscriptions")
+    group_chat = relationship("GroupChat", backref=backref("subscriptions", lazy="dynamic"))
+
+    def muted_display(self):
+        """
+        Returns (muted, muted_until) display values:
+        1. If not muted, returns (False, None)
+        2. If muted forever, returns (True, None)
+        3. If muted until a given datetime returns (True, dt)
+        """
+        if self.muted_until < now():
+            return (False, None)
+        elif self.muted_until == DATETIME_INFINITY:
+            return (True, None)
+        else:
+            return (True, self.muted_until)
+
+    @hybrid_property
+    def is_muted(self):
+        return self.muted_until > func.now()
+
+    def __repr__(self):
+        return f"GroupChatSubscription(id={self.id}, user={self.user}, joined={self.joined}, left={self.left}, role={self.role}, group_chat={self.group_chat})"
+
+
+class MessageType(enum.Enum):
+    text = enum.auto()
+    # e.g.
+    # image =
+    # emoji =
+    # ...
+    chat_created = enum.auto()
+    chat_edited = enum.auto()
+    user_invited = enum.auto()
+    user_left = enum.auto()
+    user_made_admin = enum.auto()
+    user_removed_admin = enum.auto()  # RemoveGroupChatAdmin: remove admin permission from a user in group chat
+    host_request_status_changed = enum.auto()
+    user_removed = enum.auto()  # user is removed from group chat by amdin RemoveGroupChatUser
+
+
+class Message(Base):
+    """
+    A message.
+
+    If message_type = text, then the message is a normal text message, otherwise, it's a special control message.
+    """
+
+    __tablename__ = "messages"
+
+    id = Column(BigInteger, primary_key=True)
+
+    # which conversation the message belongs in
+    conversation_id = Column(ForeignKey("conversations.id"), nullable=False, index=True)
+
+    # the user that sent the message/command
+    author_id = Column(ForeignKey("users.id"), nullable=False, index=True)
+
+    # the message type, "text" is a text message, otherwise a "control message"
+    message_type = Column(Enum(MessageType), nullable=False)
+
+    # the target if a control message and requires target, e.g. if inviting a user, the user invited is the target
+    target_id = Column(ForeignKey("users.id"), nullable=True, index=True)
+
+    # time sent, timezone should always be UTC
+    time = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
+
+    # the plain-text message text if not control
+    text = Column(String, nullable=True)
+
+    # the new host request status if the message type is host_request_status_changed
+    host_request_status_target = Column(Enum(HostRequestStatus), nullable=True)
+
+    conversation = relationship("Conversation", backref="messages", order_by="Message.time.desc()")
+    author = relationship("User", foreign_keys="Message.author_id")
+    target = relationship("User", foreign_keys="Message.target_id")
+
+    @property
+    def is_normal_message(self):
+        """
+        There's only one normal type atm, text
+        """
+        return self.message_type == MessageType.text
+
+    def __repr__(self):
+        return f"Message(id={self.id}, time={self.time}, text={self.text}, author={self.author}, conversation={self.conversation})"
