+++ app/backend/src/couchers/models/host_requests.py
@@ -1,0 +1,146 @@
+import enum
+
+from geoalchemy2 import Geometry
+from sqlalchemy import BigInteger, Boolean, Column, Date, DateTime, Enum, Float, ForeignKey, Index, String, func, text
+from sqlalchemy.ext.hybrid import hybrid_property
+from sqlalchemy.orm import column_property, relationship
+from sqlalchemy.sql import expression
+
+from couchers.models.base import Base
+from couchers.utils import date_in_timezone, now
+
+
+class HostRequestStatus(enum.Enum):
+    pending = enum.auto()
+    accepted = enum.auto()
+    rejected = enum.auto()
+    confirmed = enum.auto()
+    cancelled = enum.auto()
+
+
+class HostRequestQuality(enum.Enum):
+    high_quality = enum.auto()
+    okay_quality = enum.auto()
+    low_quality = enum.auto()
+
+
+class HostRequest(Base):
+    """
+    A request to stay with a host
+    """
+
+    __tablename__ = "host_requests"
+
+    conversation_id = Column("id", ForeignKey("conversations.id"), nullable=False, primary_key=True)
+    surfer_user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
+    host_user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
+
+    hosting_city = Column(String, nullable=False)
+    hosting_location = Column(Geometry("POINT", srid=4326), nullable=False)
+    hosting_radius = Column(Float, nullable=False)
+
+    # TODO: proper timezone handling
+    timezone = "Etc/UTC"
+
+    # dates in the timezone above
+    from_date = Column(Date, nullable=False)
+    to_date = Column(Date, nullable=False)
+
+    # timezone aware start and end times of the request, can be compared to now()
+    start_time = column_property(date_in_timezone(from_date, timezone))
+    end_time = column_property(date_in_timezone(to_date, timezone) + text("interval '1 days'"))
+    start_time_to_write_reference = column_property(date_in_timezone(to_date, timezone))
+    # notice 1 day for midnight at the *end of the day*, then 14 days to write a ref
+    end_time_to_write_reference = column_property(date_in_timezone(to_date, timezone) + text("interval '15 days'"))
+
+    status = Column(Enum(HostRequestStatus), nullable=False)
+    is_host_archived = Column(Boolean, nullable=False, default=False, server_default=expression.false())
+    is_surfer_archived = Column(Boolean, nullable=False, default=False, server_default=expression.false())
+
+    host_last_seen_message_id = Column(BigInteger, nullable=False, default=0)
+    surfer_last_seen_message_id = Column(BigInteger, nullable=False, default=0)
+
+    # number of reference reminders sent out
+    host_sent_reference_reminders = Column(BigInteger, nullable=False, server_default=text("0"))
+    surfer_sent_reference_reminders = Column(BigInteger, nullable=False, server_default=text("0"))
+    host_sent_request_reminders = Column(BigInteger, nullable=False, server_default=text("0"))
+    last_sent_request_reminder_time = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
+
+    # reason why the host/surfer marked that they didn't meet up
+    # if null then they haven't marked it such
+    host_reason_didnt_meetup = Column(String, nullable=True)
+    surfer_reason_didnt_meetup = Column(String, nullable=True)
+
+    surfer = relationship("User", backref="host_requests_sent", foreign_keys="HostRequest.surfer_user_id")
+    host = relationship("User", backref="host_requests_received", foreign_keys="HostRequest.host_user_id")
+    conversation = relationship("Conversation")
+
+    __table_args__ = (
+        # allows fast lookup as to whether they didn't meet up
+        Index(
+            "ix_host_requests_host_didnt_meetup",
+            host_reason_didnt_meetup != None,
+        ),
+        Index(
+            "ix_host_requests_surfer_didnt_meetup",
+            surfer_reason_didnt_meetup != None,
+        ),
+        # Used for figuring out who needs a reminder to respond
+        Index(
+            "ix_host_requests_status_reminder_counts",
+            status,
+            host_sent_request_reminders,
+            last_sent_request_reminder_time,
+            from_date,
+        ),
+    )
+
+    @hybrid_property
+    def can_write_reference(self):
+        return (
+            ((self.status == HostRequestStatus.confirmed) | (self.status == HostRequestStatus.accepted))
+            & (now() >= self.start_time_to_write_reference)
+            & (now() <= self.end_time_to_write_reference)
+        )
+
+    @can_write_reference.expression
+    def can_write_reference(cls):
+        return (
+            ((cls.status == HostRequestStatus.confirmed) | (cls.status == HostRequestStatus.accepted))
+            & (func.now() >= cls.start_time_to_write_reference)
+            & (func.now() <= cls.end_time_to_write_reference)
+        )
+
+    def __repr__(self):
+        return f"HostRequest(id={self.conversation_id}, surfer_user_id={self.surfer_user_id}, host_user_id={self.host_user_id}...)"
+
+
+class HostRequestFeedback(Base):
+    """
+    Private feedback from a host about a host request
+    """
+
+    __tablename__ = "host_request_feedbacks"
+
+    id = Column(BigInteger, primary_key=True)
+    time = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
+    host_request_id = Column(ForeignKey("host_requests.id"), nullable=False)
+
+    from_user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
+    to_user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
+
+    request_quality = Column(Enum(HostRequestQuality), nullable=True)
+    decline_reason = Column(String, nullable=True)  # plain text
+
+    host_request = relationship("HostRequest")
+
+    __table_args__ = (
+        # Each user can leave at most one friend reference to another user
+        Index(
+            "ix_unique_host_req_feedback",
+            from_user_id,
+            to_user_id,
+            host_request_id,
+            unique=True,
+        ),
+    )
