+++ app/backend/src/couchers/models/notifications.py
@@ -1,0 +1,296 @@
+import enum
+
+from google.protobuf import empty_pb2
+from sqlalchemy import (
+    BigInteger,
+    Boolean,
+    Column,
+    DateTime,
+    Enum,
+    ForeignKey,
+    Index,
+    Integer,
+    String,
+    UniqueConstraint,
+    func,
+)
+from sqlalchemy import LargeBinary as Binary
+from sqlalchemy.orm import relationship
+from sqlalchemy.sql import expression
+
+from couchers.constants import DATETIME_INFINITY
+from couchers.models.base import Base
+from proto import notification_data_pb2
+
+
+class NotificationDeliveryType(enum.Enum):
+    # send push notification to mobile/web
+    push = enum.auto()
+    # send individual email immediately
+    email = enum.auto()
+    # send in digest
+    digest = enum.auto()
+
+
+dt = NotificationDeliveryType
+nd = notification_data_pb2
+dt_sec = [dt.email, dt.push]
+dt_all = [dt.email, dt.push, dt.digest]
+
+
+class NotificationTopicAction(enum.Enum):
+    def __init__(self, topic_action, defaults, user_editable, data_type):
+        self.topic, self.action = topic_action.split(":")
+        self.defaults = defaults
+        # for now user editable == not a security notification
+        self.user_editable = user_editable
+
+        self.data_type = data_type
+
+    def unpack(self):
+        return self.topic, self.action
+
+    @property
+    def display(self):
+        return f"{self.topic}:{self.action}"
+
+    def __str__(self):
+        return self.display
+
+    # topic, action, default delivery types
+    friend_request__create = ("friend_request:create", dt_all, True, nd.FriendRequestCreate)
+    friend_request__accept = ("friend_request:accept", dt_all, True, nd.FriendRequestAccept)
+
+    # host requests
+    host_request__create = ("host_request:create", dt_all, True, nd.HostRequestCreate)
+    host_request__accept = ("host_request:accept", dt_all, True, nd.HostRequestAccept)
+    host_request__reject = ("host_request:reject", dt_all, True, nd.HostRequestReject)
+    host_request__confirm = ("host_request:confirm", dt_all, True, nd.HostRequestConfirm)
+    host_request__cancel = ("host_request:cancel", dt_all, True, nd.HostRequestCancel)
+    host_request__message = ("host_request:message", [dt.push, dt.digest], True, nd.HostRequestMessage)
+    host_request__missed_messages = ("host_request:missed_messages", [dt.email], True, nd.HostRequestMissedMessages)
+    host_request__reminder = ("host_request:reminder", dt_all, True, nd.HostRequestReminder)
+
+    activeness__probe = ("activeness:probe", dt_sec, False, nd.ActivenessProbe)
+
+    # you receive a friend ref
+    reference__receive_friend = ("reference:receive_friend", dt_all, True, nd.ReferenceReceiveFriend)
+    # you receive a reference from ... the host
+    reference__receive_hosted = ("reference:receive_hosted", dt_all, True, nd.ReferenceReceiveHostRequest)
+    # ... the surfer
+    reference__receive_surfed = ("reference:receive_surfed", dt_all, True, nd.ReferenceReceiveHostRequest)
+
+    # you hosted
+    reference__reminder_hosted = ("reference:reminder_hosted", dt_all, True, nd.ReferenceReminder)
+    # you surfed
+    reference__reminder_surfed = ("reference:reminder_surfed", dt_all, True, nd.ReferenceReminder)
+
+    badge__add = ("badge:add", [dt.push, dt.digest], True, nd.BadgeAdd)
+    badge__remove = ("badge:remove", [dt.push, dt.digest], True, nd.BadgeRemove)
+
+    # group chats
+    chat__message = ("chat:message", [dt.push, dt.digest], True, nd.ChatMessage)
+    chat__missed_messages = ("chat:missed_messages", [dt.email], True, nd.ChatMissedMessages)
+
+    # events
+    # approved by mods
+    event__create_approved = ("event:create_approved", dt_all, True, nd.EventCreate)
+    # any user creates any event, default to no notifications
+    event__create_any = ("event:create_any", [], True, nd.EventCreate)
+    event__update = ("event:update", dt_all, True, nd.EventUpdate)
+    event__cancel = ("event:cancel", dt_all, True, nd.EventCancel)
+    event__delete = ("event:delete", dt_all, True, nd.EventDelete)
+    event__invite_organizer = ("event:invite_organizer", dt_all, True, nd.EventInviteOrganizer)
+    event__reminder = ("event:reminder", dt_all, True, nd.EventReminder)
+    # toplevel comment on an event
+    event__comment = ("event:comment", dt_all, True, nd.EventComment)
+
+    # discussion created
+    discussion__create = ("discussion:create", [dt.digest], True, nd.DiscussionCreate)
+    # someone comments on your discussion
+    discussion__comment = ("discussion:comment", dt_all, True, nd.DiscussionComment)
+
+    # someone responds to any of your top-level comment across the platform
+    thread__reply = ("thread:reply", dt_all, True, nd.ThreadReply)
+
+    # account settings
+    password__change = ("password:change", dt_sec, False, empty_pb2.Empty)
+    email_address__change = ("email_address:change", dt_sec, False, nd.EmailAddressChange)
+    email_address__verify = ("email_address:verify", dt_sec, False, empty_pb2.Empty)
+    phone_number__change = ("phone_number:change", dt_sec, False, nd.PhoneNumberChange)
+    phone_number__verify = ("phone_number:verify", dt_sec, False, nd.PhoneNumberVerify)
+    # reset password
+    password_reset__start = ("password_reset:start", dt_sec, False, nd.PasswordResetStart)
+    password_reset__complete = ("password_reset:complete", dt_sec, False, empty_pb2.Empty)
+
+    # account deletion
+    account_deletion__start = ("account_deletion:start", dt_sec, False, nd.AccountDeletionStart)
+    # no more pushing to do
+    account_deletion__complete = ("account_deletion:complete", dt_sec, False, nd.AccountDeletionComplete)
+    # undeleted
+    account_deletion__recovered = ("account_deletion:recovered", dt_sec, False, empty_pb2.Empty)
+
+    # admin actions
+    gender__change = ("gender:change", dt_sec, False, nd.GenderChange)
+    birthdate__change = ("birthdate:change", dt_sec, False, nd.BirthdateChange)
+    api_key__create = ("api_key:create", dt_sec, False, nd.ApiKeyCreate)
+
+    donation__received = ("donation:received", dt_sec, True, nd.DonationReceived)
+
+    onboarding__reminder = ("onboarding:reminder", dt_sec, True, empty_pb2.Empty)
+
+    modnote__create = ("modnote:create", dt_sec, False, empty_pb2.Empty)
+
+    verification__sv_fail = ("verification:sv_fail", dt_sec, False, nd.VerificationSVFail)
+    verification__sv_success = ("verification:sv_success", dt_sec, False, empty_pb2.Empty)
+
+    # general announcements
+    general__new_blog_post = ("general:new_blog_post", [dt.push, dt.digest], True, nd.GeneralNewBlogPost)
+
+
+class NotificationPreference(Base):
+    __tablename__ = "notification_preferences"
+
+    id = Column(BigInteger, primary_key=True)
+    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
+
+    topic_action = Column(Enum(NotificationTopicAction), nullable=False)
+    delivery_type = Column(Enum(NotificationDeliveryType), nullable=False)
+    deliver = Column(Boolean, nullable=False)
+
+    user = relationship("User", foreign_keys="NotificationPreference.user_id")
+
+    __table_args__ = (UniqueConstraint("user_id", "topic_action", "delivery_type"),)
+
+
+class Notification(Base):
+    """
+    Table for accumulating notifications until it is time to send email digest
+    """
+
+    __tablename__ = "notifications"
+
+    id = Column(BigInteger, primary_key=True)
+    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
+
+    # recipient user id
+    user_id = Column(ForeignKey("users.id"), nullable=False)
+
+    topic_action = Column(Enum(NotificationTopicAction), nullable=False)
+    key = Column(String, nullable=False)
+
+    data = Column(Binary, nullable=False)
+
+    # whether the user has marked this notification as seen or not
+    is_seen = Column(Boolean, nullable=False, server_default=expression.false())
+
+    user = relationship("User", foreign_keys="Notification.user_id")
+
+    __table_args__ = (
+        # used in looking up which notifications need delivery
+        Index(
+            "ix_notifications_created",
+            created,
+        ),
+        # Fast lookup for unseen notification count
+        Index(
+            "ix_notifications_unseen",
+            user_id,
+            topic_action,
+            postgresql_where=(is_seen == False),
+        ),
+        # Fast lookup for latest notifications
+        Index(
+            "ix_notifications_latest",
+            user_id,
+            id.desc(),
+            topic_action,
+        ),
+    )
+
+    @property
+    def topic(self):
+        return self.topic_action.topic
+
+    @property
+    def action(self):
+        return self.topic_action.action
+
+
+class NotificationDelivery(Base):
+    __tablename__ = "notification_deliveries"
+
+    id = Column(BigInteger, primary_key=True)
+    notification_id = Column(ForeignKey("notifications.id"), nullable=False, index=True)
+    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
+    delivered = Column(DateTime(timezone=True), nullable=True)
+    read = Column(DateTime(timezone=True), nullable=True)
+    # todo: enum of "phone, web, digest"
+    delivery_type = Column(Enum(NotificationDeliveryType), nullable=False)
+    # todo: device id
+    # todo: receipt id, etc
+    notification = relationship("Notification", foreign_keys="NotificationDelivery.notification_id")
+
+    __table_args__ = (
+        UniqueConstraint("notification_id", "delivery_type"),
+        # used in looking up which notifications need delivery
+        Index(
+            "ix_notification_deliveries_delivery_type",
+            delivery_type,
+            postgresql_where=(delivered != None),
+        ),
+        Index(
+            "ix_notification_deliveries_dt_ni_dnull",
+            delivery_type,
+            notification_id,
+            delivered == None,
+        ),
+    )
+
+
+class PushNotificationSubscription(Base):
+    __tablename__ = "push_notification_subscriptions"
+
+    id = Column(BigInteger, primary_key=True)
+    created = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
+
+    # which user this is connected to
+    user_id = Column(ForeignKey("users.id"), nullable=False, index=True)
+
+    # these come from https://developer.mozilla.org/en-US/docs/Web/API/PushSubscription
+    # the endpoint
+    endpoint = Column(String, nullable=False)
+    # the "auth" key
+    auth_key = Column(Binary, nullable=False)
+    # the "p256dh" key
+    p256dh_key = Column(Binary, nullable=False)
+
+    full_subscription_info = Column(String, nullable=False)
+
+    # the browse user-agent, so we can tell the user what browser notifications are going to
+    user_agent = Column(String, nullable=True)
+
+    # when it was disabled
+    disabled_at = Column(DateTime(timezone=True), nullable=False, server_default=DATETIME_INFINITY.isoformat())
+
+    user = relationship("User")
+
+
+class PushNotificationDeliveryAttempt(Base):
+    __tablename__ = "push_notification_delivery_attempt"
+
+    id = Column(BigInteger, primary_key=True)
+    time = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
+
+    push_notification_subscription_id = Column(
+        ForeignKey("push_notification_subscriptions.id"), nullable=False, index=True
+    )
+
+    success = Column(Boolean, nullable=False)
+    # the HTTP status code, 201 is success
+    status_code = Column(Integer, nullable=False)
+
+    # can be null if it was a success
+    response = Column(String, nullable=True)
+
+    push_notification_subscription = relationship("PushNotificationSubscription")
